Module theseeker_engine::prelude

Re-exports

pub use crate::assets::AssetKey;

pub use crate::assets::AssetsSet;

pub use crate::assets::PreloadedAssets;

pub use crate::data::TickQuant;

pub use crate::time::at_tick_multiples;

pub use crate::time::GameTickMidFlush;

pub use crate::time::GameTickSet;

pub use crate::time::GameTickUpdate;

pub use crate::time::GameTime;

Modules

broad_phase

Collects pairs of potentially colliding entities into BroadCollisionPairs using AABB intersection checks.

helpers

A module which provides helper functions.

integrator

Integrates Newton’s 2nd law of motion, applying forces and moving entities according to their velocities.

joints

Contains joints.

ldtk

Contains all the types for serializing/deserializing an LDtk file.

penetration

Penetration constraint.

prepare

Runs systems at the start of each physics frame; initializes rigid bodies and colliders and updates components.

setup

Sets up the physics engine by initializing the necessary schedules, sets and resources.

shape

Generation for some primitive shape meshes.

sleeping

Controls when bodies should be deactivated and marked as Sleeping to improve performance.

solver

Solves positional and angular constraints, updates velocities and solves velocity constraints (dynamic friction, restitution and joint damping).

spatial_query

Spatial queries are a way to get information about the environment. They perform geometric queries on colliders and retrieve data about intersections.

sync

Responsible for synchronizing physics components with other data, like writing Position and Rotation components to Transforms.

system_adapter

A collection of common adapters for piping the result of a system.

Macros

anyhow

Construct an ad-hoc error from a string or existing non-anyhow error value.

bail

Return early with an error.

debug

Constructs an event at the debug level.

debug_span

Constructs a span at the debug level.

ensure

error

Constructs an event at the error level.

error_span

Constructs a span at the error level.

info

Constructs an event at the info level.

info_span

Constructs a span at the info level.

trace

Constructs an event at the trace level.

trace_span

Constructs a span at the trace level.

warn

Constructs an event at the warn level.

warn_span

Constructs a span at the warn level.

Structs

AabbGizmo

Add this Component to an entity to draw its Aabb component.

AabbGizmoConfig

Configuration for drawing the Aabb component on entities.

AddChild

Command that adds a child to an entity.

Added

A filter on a component that only retains results added after the system last ran.

AncestorIter

An Iterator of Entitys over the ancestors of an Entity.

AngleLimit

A limit that indicates that angles should be between alpha and beta.

AngularDamping

Automatically slows down a dynamic rigid body, decreasing it’s angular velocity each frame. This can be used to simulate air resistance.

AngularVelocity

The angular velocity of a body in radians. Positive values will result in counterclockwise rotation.

AnimatedTile

A component that is attached to a Tile entity that tells the GPU how to animate the tile. Currently all frames must be aligned in your tilemap.

AnimationClip

A list of VariableCurve, and the EntityPath to which they apply.

AnimationPlayer

Animation controls

AnimationPlugin

Adds animation support to an app

AnyError

The Error type, a wrapper around a dynamic error type.

AnyOf

The AnyOf query parameter fetches entities with any of the component types included in T.

App

A container of app logic and data.

AppTypeRegistry

A Resource storing TypeRegistry for type registrations relevant to a whole app.

ArrayTextureLoader

A bevy world resource that allows you to add atlas textures for loading into the array texture cache.

AssetPlugin

Adds support for Assets to an App.

AssetServer

Loads assets from the filesystem in the background.

Assets

Stores Assets of a given type and tracks changes to them.

AssetsLoading

Resource for tracking the loading of assets

AssetsTrackProgress

System Set for assets progress tracking.

AtlasImageBundle

A UI node that is a texture atlas sprite

AudioSink

Used to control audio during playback.

AudioSource

A source of audio data

AudioSourceBundle

Bundle for playing a sound.

Axis

Stores the position data of the input devices of type T.

BSpline

A spline interpolated continuously across the nearest four control points. The curve does not pass through any of the control points.

BVec2

A 2-dimensional bool vector mask.

BVec3

A 3-dimensional bool vector mask.

BVec4

A 4-dimensional bool vector mask.

BackgroundColor

The background color of the node

Bezier

A spline composed of a series of cubic Bezier curves.

BorderColor

The border color of the UI node.

BroadCollisionPairs

A list of entity pairs for potential collisions collected during the broad phase.

BroadPhasePlugin

Collects pairs of potentially colliding entities into BroadCollisionPairs using AABB intersection checks. This speeds up narrow phase collision detection, as the number of precise collision checks required is greatly reduced.

Button

Marker struct for buttons

ButtonBundle

A UI node that is a button

CalculatedClip

The calculated clip of the node

Camera

The defining Component for camera entities, storing information about how and what to render through this camera.

Camera2d

Camera2dBundle

Camera3d

Configuration for the “main 3d render graph”.

Camera3dBundle

CardinalSpline

A spline interpolated continuously across the nearest four control points, with the position of the curve specified at every control point and the tangents computed automatically.

CenterOfMass

The local center of mass of a body.

ChainOptionalSystem

Similar to Bevy’s PipedSystem, but the second system is optional

ChainResultSystem

Similar to Bevy’s PipedSystem, but with diverging paths for Ok/Err Results

Changed

A filter on a component that only retains results added or mutably dereferenced after the system last ran.

ChildBuilder

Struct for building children entities and adding them to a parent entity.

Children

Contains references to the child entities of this entity.

ClearChildren

Command that clears all children from an entity and removes Parent component from those children.

ClearColor

A Resource that stores the color that is used to clear the screen between frames.

Collider

A collider used for collision detection.

ColliderAabb

The Axis-Aligned Bounding Box of a collider.

ColliderMassProperties

The mass properties derived from a given collider shape and density.

CollidingEntities

Contains the entities that are colliding with an entity.

Collision

A collision event that is sent for each contact pair during the narrow phase.

CollisionEnded

A collision event that is sent when two entities stop colliding.

CollisionLayers

Defines the collision layers of a collider using groups and masks.

CollisionStarted

A collision event that is sent when two entities start colliding.

ColorMaterial

A 2d material that renders 2d meshes with a texture tinted by a uniform color

Commands

A Command queue to perform impactful changes to the World.

ComputedVisibility

Algorithmically-computed indication of whether an entity is visible and should be extracted for rendering

Contact

Data related to a contact between two bodies.

CubicSegment

A segment of a cubic curve, used to hold precomputed coefficients for fast interpolation.

CursorEntered

An event that is sent whenever the user’s cursor enters a window.

CursorLeft

An event that is sent whenever the user’s cursor leaves a window.

CursorMoved

An event reporting that the mouse cursor has moved inside a window.

DeactivationTime

How long in seconds the linear and angular velocity of a body need to be below the SleepingThreshold before the body is deactivated. Defaults to 1 second.

DebugName

Convenient query for giving a human friendly name to an entity.

DefaultPlugins

This plugin group will add all the default plugins for a Bevy application:

Deferred

A SystemParam that stores a buffer which gets applied to the World during apply_deferred. This is used internally by Commands to defer World mutations.

DeltaTime

How much time the previous physics frame took. The timestep can be configured with the PhysicsTimestep resource.

DescendantIter

An Iterator of Entitys over the descendants of an Entity.

DespawnChildrenRecursive

Despawns the given entity’s children recursively

DespawnRecursive

Despawns the given entity and all its children recursively

DistanceLimit

A limit that indicates that the distance between two points should be between min and max.

Duration

A Duration type to represent a span of time, typically used for system timeouts.

DynamicAssetCollections

Resource keeping track of dynamic asset collection files for different loading states

DynamicAssets

Resource to dynamically resolve keys to assets.

DynamicScene

A collection of serializable resources and dynamic entities.

DynamicSceneBuilder

A DynamicScene builder, used to build a scene from a World by extracting some entities and resources.

DynamicSceneBundle

A component bundle for a DynamicScene root.

Entity

Lightweight identifier of an entity.

EntityIid

Component added to all LdtkEntitys by default.

EntityInstance

Component added to any LDtk Entity by default.

EntityPath

Path to an entity, with Names. Each entity in a path must have a name.

EntityRef

A read-only reference to a particular Entity and all of its components

EventReader

Reads events of type T in order and tracks which events have already been read.

EventWriter

Sends events of type T.

Events

An event collection that represents the events that occurred within the last two Events::update calls. Events can be written to using an EventWriter and are typically cheaply read using an EventReader.

ExternalForce

An external force applied to a dynamic rigid body during the integration step.

ExternalTorque

An external torque applied to a dynamic rigid body during the integration step.

ExtractSchedule

Schedule which extract data from the main world and inserts it into the render world.

First

Runs first in the schedule. This is run by the Main schedule.

FixedJoint

A fixed joint prevents any relative movement of the attached bodies.

FixedTime

The amount of time that must pass before the fixed timestep schedule is run again.

FixedUpdate

The schedule that contains systems which only run after a fixed period of time has elapsed.

Font

FrameCountPlugin

Adds frame counting functionality to Apps.

Friction

Controls how strongly the material of an entity prevents relative tangential movement at contact points.

Gamepad

A gamepad with an associated ID.

GamepadAxis

An axis of a Gamepad.

GamepadButton

A button of a Gamepad.

Gamepads

A collection of connected Gamepads.

GilrsPlugin

GizmoConfig

A Resource that stores configuration for gizmos.

Gizmos

A SystemParam for drawing gizmos.

GlobalTransform

Describe the position of an entity relative to the reference frame.

GlobalVolume

Use this Resource to control the global volume of all audio with a Volume::Relative volume.

Gravity

A resource for the global gravitational acceleration.

GravityScale

Controls how gravity affects a specific rigid body.

GridCoords

Component that stores grid-based coordinate information.

GridPlacement

Represents the position of a grid item in a single axis.

GridTrack

A GridTrack is a Row or Column of a CSS Grid. This struct specifies what size the track should be. See below for the different “track sizing functions” you can specify.

Handle

A handle into a specific Asset of type T.

HandleUntyped

A non-generic version of Handle.

Hermite

A spline interpolated continuously between the nearest two control points, with the position and velocity of the curve specified at both control points. This curve passes through all control points, with the specified velocity which includes direction and parametric speed.

HiddenProgress

“Hidden” progress reported by a system.

HierarchyPlugin

The base plugin for handling Parent and Children components

IVec2

A 2-dimensional vector.

IVec3

A 3-dimensional vector.

IVec4

A 4-dimensional vector.

Image

ImageBundle

A UI node that is an image

ImagePlugin

Adds the Image as an asset and makes sure that they are extracted and prepared for the GPU.

In

Wrapper type to mark a SystemParam as an input.

Inertia

The moment of inertia of a body. This represents the torque needed for a desired angular acceleration.

Input

A “press-able” input of type T.

InsertChildren

Command that inserts a child at a given index of a parent’s children, shifting following children back.

IntGridCell

Component added to any IntGrid tile by default.

IntegratorPlugin

Integrates Newton’s 2nd law of motion, applying forces and moving entities according to their velocities.

InverseInertia

The inverse moment of inertia of the body. This represents the inverse of the torque needed for a desired angular acceleration.

InverseMass

The inverse mass of a body.

IterationCount

The number of iterations used in the position solver. It is recommended to keep this low and to increase SubstepCount instead, as substepping can provide better convergence, accuracy and energy conservation.

Label

Marker struct for labels

Last

Runs last in the schedule. This is run by the Main schedule.

LayerInstance

LayerMetadata

Component for storing some LDtk layer information on layer entities.

LdtkAsset

Main asset for loading ldtk files.

LdtkLevel

Secondary asset for loading ldtk files, specific to level data.

LdtkPlugin

Adds the default systems, assets, and resources used by bevy_ecs_ldtk.

LdtkSettings

Settings resource for the plugin. Check out the documentation for each field type to learn more.

LdtkWorldBundle

Bundle for spawning LDtk worlds and their levels. The main bundle for using this plugin.

LevelSet

Component that determines the desired levels to be loaded for an LdtkWorldBundle.

LinearDamping

Automatically slows down a dynamic rigid body, decreasing it’s linear velocity each frame. This can be used to simulate air resistance.

LinearVelocity

The linear velocity of a body.

LoadingState

A Bevy plugin to configure automatic asset loading

Local

A system local SystemParam.

LockedAxes

A component that specifies which translational and rotational axes of a rigid body are locked.

Main

The schedule that contains the app logic that is evaluated each tick of App::update().

Mass

The mass of a body.

MassPropertiesBundle

A bundle containing mass properties.

Mat2

A 2x2 column major matrix.

Mat3

A 3x3 column major matrix.

Mat4

A 4x4 column major matrix.

MaterialTilemapBundle

The default tilemap bundle. All of the components within are required.

MaterialTilemapPlugin

Mesh

A 3D object made out of vertices representing triangles, lines, or points, with “attribute” values for each vertex.

MinimalPlugins

This plugin group will add the minimal plugins for a Bevy application:

MorphWeights

Controls the morph targets for all child Handle<Mesh> entities. In most cases, MorphWeights should be considered the “source of truth” when writing morph targets for meshes. However you can choose to write child MeshMorphWeights if your situation requires more granularity. Just note that if you set MorphWeights, it will overwrite child MeshMorphWeights values.

Mut

Unique mutable borrow of an entity’s component

Name

Component used to identify an entity. Stores a hash for faster comparisons.

NextState

The next state of State<S>.

Node

Describes the size of a UI node

NodeBundle

The basic UI node

NonSend

Shared borrow of a non-Send resource.

NonSendMut

Unique borrow of a non-Send resource.

OnClick

Add this component to an interactable UI node, to specify things to run automatically on click.

OnEnter

The label of a Schedule that runs whenever State<S> enters this state.

OnExit

The label of a Schedule that runs whenever State<S> exits this state.

OnTransition

The label of a Schedule that only runs whenever State<S> exits the from state, AND enters the to state.

Or

A filter that tests if any of the given filters apply.

OrthographicProjection

Project a 3D space onto a 2D surface using parallel lines, i.e., unlike PerspectiveProjection, the size of objects remains the same regardless of their distance to the camera.

Overflow

Whether to show or hide overflowing items

ParallelCommands

An alternative to Commands that can be used in parallel contexts, such as those in Query::par_iter

ParamSet

A collection of potentially conflicting SystemParams allowed by disjoint access.

Parent

Holds a reference to the parent entity of this entity. This component should only be present on entities that actually have a parent entity.

PenetrationConstraint

A constraint between two bodies that prevents overlap with a given compliance.

PerspectiveProjection

A 3D camera projection in which distant objects appear smaller than close objects.

PhysicsLoop

Data related to the physics simulation loop.

PhysicsPlugins

A plugin group containing all of Bevy XPBD’s plugins.

PhysicsSetupPlugin

Sets up the physics engine by initializing the necessary schedules, sets and resources.

PlaybackSettings

Initial settings to be used when audio starts playing. If you would like to control the audio while it is playing, query for the AudioSink or SpatialAudioSink components. Changes to this component will not be applied to already-playing audio.

PointProjection

The result of a point projection on a collider.

Position

The position of a body.

PostStartup

The schedule that runs once after Startup. This is run by the Main schedule.

PostUpdate

The schedule that contains logic that must run after Update. For example, synchronizing “local transforms” in a hierarchy to “global” absolute transforms. This enables the PostUpdate transform-sync system to react to “local transform” changes in Update without the Update systems needing to know about (or add scheduler dependencies for) the “global transform sync system”.

PreStartup

The schedule that runs before Startup. This is run by the Main schedule.

PreUpdate

The schedule that contains logic that must run before Update. For example, a system that reads raw keyboard input OS events into an Events resource. This enables systems in Update to consume the events from the Events resource without actually knowing about (or taking a direct scheduler dependency on) the “os-level keyboard event sytsem”.

PreparePlugin

Runs systems at the start of each physics frame; initializes rigid bodies and colliders and updates components.

PreviousPosition

The previous position of a body.

PreviousRotation

The previous rotation of a body. See Rotation.

PrismaticJoint

A prismatic joint prevents relative movement of the attached bodies, except for translation along one free_axis.

ProcessLdtkApi

Schedule for processing this plugin’s ECS API, inserted after Update.

Progress

Progress reported by a system

ProgressCounter

Resource for tracking overall progress

ProgressPlugin

Add this plugin to your app, to use this crate for the specified state.

PushChildren

Command that pushes children to the end of the entity’s Children.

Quat

A quaternion representing an orientation.

Query

System parameter that provides selective access to the Component data stored in a World.

QueryState

Provides scoped access to a World state according to a given WorldQuery and query filter.

Ray

A ray is an infinite line starting at origin, going in direction.

RayCaster

A component used for ray casting.

RayHitData

Data related to a hit during a ray cast.

RayHits

Contains the hits of a ray cast by a RayCaster.

ReceivedCharacter

An event that is sent whenever a window receives a character from the OS or underlying system.

Rect

A rectangle defined by two opposite corners.

Ref

Shared borrow of an entity’s component with access to change detection. Similar to Mut but is immutable and so doesn’t require unique access.

ReflectComponent

A struct used to operate on reflected Component of a type.

ReflectDefault

A struct used to provide the default value of a type.

ReflectDeserialize

A struct used to deserialize reflected instances of a type.

ReflectFromReflect

Type data that represents the FromReflect trait and allows it to be used dynamically.

ReflectResource

A struct used to operate on reflected Resource of a type.

ReflectSerialize

A struct used to serialize reflected instances of a type.

RegisterStandardDynamicAsset

Command to register a standard dynamic asset under the given key

RemoveChildren

Command that removes children from an entity, and removes these children’s parent.

RemoveParent

Command that removes the parent of an entity, and removes that entity from the parent’s Children.

RemovedComponents

A SystemParam that grants access to the entities that had their T Component removed.

RepeatedGridTrack

Represents a possibly repeated GridTrack.

ReplaceChildren

Command that clear all children from an entity, replacing them with the given children.

Res

Shared borrow of a Resource.

ResMut

Unique mutable borrow of a Resource.

Respawn

Component that indicates that an LDtk level or world should respawn.

Restitution

Controls how elastic or bouncy an entity is when colliding with other entities.

RevoluteJoint

A revolute joint prevents relative movement of the attached bodies, except for rotation around one aligned_axis.

RigidBodyQuery

A WorldQuery to make querying and modifying rigid bodies more convenient.

RigidBodyQueryItem

Automatically generated WorldQuery item type for RigidBodyQuery, returned when iterating over query results.

RigidBodyQueryReadOnly

Automatically generated WorldQuery type for a read-only variant of RigidBodyQuery.

RigidBodyQueryReadOnlyItem

Automatically generated WorldQuery item type for RigidBodyQueryReadOnly, returned when iterating over query results.

Rotation

The rotation of a body.

RumbleSystem

Updates the running gamepad rumble effects.

ScanCode

The scan code of a KeyboardInput.

Scene

To spawn a scene, you can use either:

SceneBundle

A component bundle for a Scene root.

SceneSpawner

Schedule

A collection of systems, and the metadata and executor needed to run them in a certain order under certain conditions.

Schedules

Resource that stores Schedules mapped to ScheduleLabels.

Sensor

A component that marks a Collider as a sensor collider.

Shader

A shader, as defined by its ShaderSource and ShaderStage This is an “unprocessed” shader. It can contain preprocessor directives.

ShapeCaster

A component used for shape casting.

ShapeHitData

Data related to a hit during a shape cast.

ShapeHits

Contains the hits of a shape cast by a ShapeCaster. The hits are in the order of time of impact.

Sleeping

Indicates that a body is not simulated by the physics engine until woken up again. This is done to improve performance and to help prevent small jitter that is typically present in collisions.

SleepingDisabled

Indicates that the body can not be deactivated by the physics engine. See Sleeping for information about sleeping.

SleepingPlugin

Controls when bodies should be deactivated and marked as Sleeping to improve performance.

SleepingThreshold

A threshold that indicates the maximum linear and angular velocity allowed for a body to be deactivated.

SolverPlugin

Solves positional and angular constraints, updates velocities and solves velocity constraints (dynamic friction and restitution and joint damping).

SpatialAudioSink

Used to control spatial audio during playback.

SpatialAudioSourceBundle

Bundle for playing a sound with a 3D position.

SpatialBundle

A Bundle with the following Components:

SpatialQuery

A system parameter for performing spatial queries.

SpatialQueryFilter

Rules that determine which colliders are taken into account in spatial queries.

SpatialQueryPipeline

A resource for the spatial query pipeline.

SpatialQueryPlugin

Initializes the SpatialQueryPipeline resource and handles component-based spatial queries like ray casting and shape casting with RayCaster and ShapeCaster.

SpatialSettings

Settings for playing spatial audio.

SphericalJoint

A spherical joint prevents relative translation of the attached bodies while allowing rotation around all axes.

Sprite

SpriteBundle

SpriteSheetBundle

A Bundle of components for drawing a single sprite from a sprite sheet (also referred to as a TextureAtlas)

StandardDynamicAssetCollection

The asset defining a mapping from asset keys to dynamic assets

StandardTilemapMaterial

Startup

The schedule that runs once when the app starts. This is run by the Main schedule.

State

A finite-state machine whose transitions have associated schedules ([OnEnter(state)] and [OnExit(state)]).

StateTransition

Runs state transitions. This is run by the Main schedule.

StdRng

The standard RNG. The PRNG algorithm in StdRng is chosen to be efficient on the current platform, to be statistically strong and unpredictable (meaning a cryptographically secure PRNG).

Style

Describes the style of a UI container node

SubDeltaTime

How much time the previous physics substep took. This depends on the DeltaTime and SubstepCount resources.

SubstepCount

The number of substeps used in the simulation.

SyncPlugin

Responsible for synchronizing physics components with other data, like writing Position and Rotation components to Transforms.

TaskPoolOptions

Helper for configuring and creating the default task pools. For end-users who want full control, set up TaskPoolPlugin

TaskPoolPlugin

Setup of default task pools: AsyncComputeTaskPool, ComputeTaskPool, IoTaskPool.

Text

Text2dBundle

The bundle of components needed to draw text in a 2D scene via a 2D Camera2dBundle. Example usage.

TextBundle

A UI node that is text

TextSection

TextStyle

TextureAtlas

An atlas containing multiple textures (like a spritesheet or a tilemap). Example usage animating sprite. Example usage loading sprite sheet.

TextureAtlasBuilder

A builder which is used to create a texture atlas from many individual sprites.

TextureAtlasSprite

ThreadRng

A reference to the thread-local generator

TileBundle

This an optional tile bundle with default components.

TileColor

A custom color for the tile.

TileEnumTags

Component for storing user-defined, enum-based tags for a particular tile in an LDtk tileset definition.

TileFlip

Flips the tiles texture along the X, Y or diagonal axes

TileMetadata

Component for storing user-defined custom data for a paticular tile in an LDtk tileset definition.

TilePos

A tile position in the tilemap grid.

TilePosOld

TileStorage

Used to store tile entities for fast look up. Tile entities are stored in a grid. The grid is always filled with None.

TileTextureIndex

A texture index into the atlas or texture array for a single tile. Indices in an atlas are horizontal based.

TileVisible

Hides or shows a tile based on the boolean. Default: True

TilemapArrayTexture

TilemapGridSize

Size of the tiles on the grid in pixels. This can be used to overlay tiles on top of each other. Ex. A 16x16 pixel tile can be overlapped by 8 pixels by using a grid size of 16x8.

TilemapId

A component which stores a reference to the tilemap entity.

TilemapPlugin

A bevy tilemap plugin. This must be included in order for everything to be rendered. But is not necessary if you are running without a renderer.

TilemapRenderSettings

Custom parameters for the render pipeline.

TilemapSize

Size of the tilemap in tiles.

TilemapSpacing

Spacing between tiles in pixels inside of the texture atlas. Defaults to 0.0

TilemapTextureSize

Size of the atlas texture in pixels.

TilemapTileSize

Size of the tiles in pixels

TilesetDefinition

The Tileset definition is the most important part among project definitions. It contains some extra informations about each integrated tileset. If you only had to parse one definition section, that would be the one.

Time

A clock that tracks how much it has advanced (and how much real time has elapsed) since its previous update and since its creation.

TimeSleeping

How long the velocity of the body has been below the SleepingThreshold, i.e. how long the body has been able to sleep.

Timer

Tracks elapsed time. Enters the finished state once duration is reached.

TouchInput

A touch input event.

Touches

A collection of Touches.

Transform

Describe the position of an entity. If the entity has a parent, the position is relative to its parent position.

TransformBundle

A Bundle of the Transform and GlobalTransform Components, which describe the position of an entity.

TransformPlugin

The base plugin for handling Transform components

TypeRegistrationPlugin

Registration of default types to the TypeRegistry resource.

UVec2

A 2-dimensional vector.

UVec3

A 3-dimensional vector.

UVec4

A 4-dimensional vector.

UiCameraConfig

Configuration for cameras related to UI.

UiDisabled

For disabling UI elements; our systems for handling Interactions should look for this

UiExtrasPlugin

UiImage

The 2D texture displayed for this UI node

UiRect

A type which is commonly used to define margins, paddings and borders.

UiScale

The current scale of the UI.

UiTextureAtlasImage

The atlas sprite to be used in a UI Texture Atlas Node

Update

The schedule that contains app logic. This is run by the Main schedule.

ValidParentCheckPlugin

Print a warning for each Entity with a T component whose parent doesn’t have a T component.

VariableCurve

Describes how an attribute of a Transform or MorphWeights should be animated.

Vec2

A 2-dimensional vector.

Vec3

A 3-dimensional vector.

Vec4

A 4-dimensional vector.

VisibilityBundle

A Bundle of the Visibility and ComputedVisibility Components, which describe the visibility of an entity.

Window

The defining Component for window entities, storing information about how it should appear and behave.

WindowMoved

An event that is sent when a window is repositioned in physical pixels.

WindowPlugin

A Plugin that defines an interface for windowing support in Bevy.

WindowResizeConstraints

The size limits on a Window.

With

Filter that selects entities with a component T.

Without

Filter that selects entities without a component T.

World

Stores and exposes operations on entities, components, resources, and their associated metadata.

WorldChildBuilder

Struct for adding children to an entity directly through the World for use in exclusive systems.

WorldCursor

WorldCursorCamera

WorldCursorPlugin

WorldCursorSet

Worldly

Component that indicates that an ldtk entity should be a child of the world, not the level.

Enums

AlignContent

Defines how each line is aligned within the flexbox.

AlignItems

How items are aligned according to the cross axis

AlignSelf

How this item is aligned according to the cross axis. Overrides AlignItems.

AssetEvent

Events that involve assets of type T.

CoefficientCombine

Determines how coefficients are combined. The default is Average.

Color

CursorIcon

The icon to display for a Window’s Cursor.

Direction

Defines the text direction

Display

Whether to use a Flexbox layout model.

DynamicAssetType

Different typed that can generate the asset field value of a dynamic asset

EulerRot

Euler rotation sequences.

FieldValue

The actual value of a field instance on a Level or EntityInstance.

FileDragAndDrop

Events related to files being dragged and dropped on a window.

FlexDirection

Defines how flexbox items are ordered within a flexbox

FlexWrap

Defines if flexbox items appear on a single line or on multiple lines

GamepadAxisType

A type of a GamepadAxis.

GamepadButtonType

A type of a GamepadButton.

GridAutoFlow

Controls whether grid items are placed row-wise or column-wise. And whether the sparse or dense packing algorithm is used.

GridTrackRepetition

How many times to repeat a repeated grid track

HexCoordSystem

Different hex grid coordinate systems. You can find out more at this link: https://www.redblobgames.com/grids/hexagons/

Ime

A Input Method Editor event.

IntGridRendering

Option in LdtkSettings that determines the visual representation of IntGrid layers when they don’t have AutoTile rules.

Interaction

Describes what type of input interaction has occurred for a UI node.

IsoCoordSystem

Different isometric coordinate systems.

JustifyContent

Defines how items are aligned according to the main axis

JustifyItems

How items are aligned according to the cross axis

JustifySelf

How this item is aligned according to the cross axis. Overrides AlignItems.

KeyCode

The key code of a KeyboardInput.

Keyframes

List of keyframes for one of the attribute of a Transform.

LevelBackground

Option in LdtkSettings that dictates how the plugin handles level backgrounds.

LevelEvent

Events fired by the plugin related to level spawning/despawning.

LevelSelection

Resource for choosing which level(s) to spawn.

LevelSpawnBehavior

Option in LdtkSettings that determines level spawn behavior.

MaxTrackSizingFunction

MinTrackSizingFunction

MonitorSelection

References a screen monitor.

MouseButton

A button on a mouse device.

Msaa

Configuration resource for Multi-Sample Anti-Aliasing.

OverflowAxis

Whether to show or hide overflowing items

PhysicsSet

High-level system sets for the main phases of the physics engine. You can use these to schedule your own systems before or after physics is run without having to worry about implementation details.

PhysicsTimestep

Configures how many times per second the physics simulation is run.

PositionType

The strategy used to position this node

Projection

A configurable CameraProjection that can select its projection type at runtime.

RigidBody

A non-deformable body used for the simulation of most physics objects.

SceneFilter

A filter used to control which types can be added to a DynamicScene.

SetClearColor

Option in LdtkSettings that determines clear color behavior.

StandardDynamicAsset

These asset variants can be loaded from configuration files. They will then replace a dynamic asset based on their keys.

TextAlignment

Describes horizontal alignment preference for positioning & bounds.

TextError

TilemapTexture

TilemapType

The type of tile to be rendered, currently we support: Square, Hex, and Isometric.

TimerMode

Specifies Timer behavior.

Val

Represents the possible value types for layout properties.

ValArithmeticError

Visibility

User indication of whether an entity is visible. Propagates down the entity hierarchy.

WindowPosition

Defines where a Window should be placed on the screen.

ZIndex

Indicates that this Node entity’s front-to-back ordering is not controlled solely by its location in the UI hierarchy. A node with a higher z-index will appear on top of other nodes with a lower z-index.

Traits

AddAsset

App extension methods for adding new asset types.

AngularConstraint

An angular constraint applies an angular correction around a given axis.

AssetCollection

Trait to mark a struct as a collection of assets

AssetCollectionApp

Extension trait for App enabling initialisation of asset collections

AssetCollectionWorld

Extension trait for World enabling initialisation of asset collections

AudioSinkPlayback

Common interactions with an audio sink.

BuildChildren

Trait for removing, adding and replacing children and parents of an entity.

BuildChildrenTransformExt

Collection of methods similar to [BuildChildren], but preserving each entity’s GlobalTransform.

BuildWorldChildren

Trait that defines adding, changing and children and parents of an entity directly through the World.

Bundle

The Bundle trait enables insertion and removal of Components from an entity.

CliCommandsRegisterExt

Provides methods for managing the available “console commands”

CliCommandsRunExt

Provides methods to run/call “console commands”

Component

A data type that can be used to store data for an entity.

Condition

A system that determines if one or more scheduled systems should run.

Context

Provides the context method for Result.

CryptoRng

A marker trait used to indicate that an RngCore or BlockRngCore implementation is supposed to be cryptographically secure.

CubicGenerator

Implement this on cubic splines that can generate a curve from their spline parameters.

Decodable

A type implementing this trait can be converted to a rodio::Source type. It must be Send and Sync, and usually implements Asset so needs to be TypeUuid, in order to be registered. Types that implement this trait usually contain raw sound data that can be converted into an iterator of samples. This trait is implemented for AudioSource. Check the example audio/decodable for how to implement this trait on a custom type.

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

DeserializeOwned

A data structure that can be deserialized without borrowing any data from the deserializer.

DespawnRecursiveExt

Trait that holds functions for despawning recursively down the transform hierarchy

DetectChanges

Types that can read change detection information. This change detection is controlled by DetectChangesMut types such as ResMut.

DetectChangesMut

Types that implement reliable change detection.

Distribution

Types (distributions) that can be used to create a random instance of T.

DynamicAsset

Any type implementing this trait can be assigned to asset keys as part of a dynamic asset collection.

DynamicAssetCollection

This traits describes types that contain asset configurations and can register them in the DynamicAssets resource.

Event

A type that can be stored in an Events<E> resource You can conveniently access events using the EventReader and EventWriter system parameter.

FromReflect

A trait that enables types to be dynamically constructed from reflected data.

FromWorld

Creates an instance of the type this trait is implemented for using data from the supplied World.

GetField

A convenience trait which combines fetching and downcasting of struct fields.

GetTupleStructField

A convenience trait which combines fetching and downcasting of tuple struct fields.

HierarchyQueryExt

An extension trait for Query that adds hierarchy related methods.

IntoChainOptionalSystem

IntoChainResultSystem

IntoSystem

Conversion trait to turn something into a System.

IntoSystemConfigs

Types that can convert into a SystemConfigs.

IntoSystemSet

Types that can be converted into a SystemSet.

IntoSystemSetConfig

Types that can be converted into a SystemSetConfig.

IntoSystemSetConfigs

Types that can convert into a SystemSetConfigs.

IteratorRandom

Extension trait on iterators, providing random sampling methods.

Joint

A trait for joints.

LdtkEntity

Provides a constructor which can be used for spawning entities from an LDtk file.

LdtkEntityAppExt

Provides functions to register Bundles to bevy’s App for particular LDtk layer and entity identifiers.

LdtkFields

Convenience methods for accessing field instances.

LdtkIntCell

Provides a constructor which can be used for spawning additional components on IntGrid tiles.

LdtkIntCellAppExt

Provides functions to register Bundles to bevy’s App for particular LDtk layer and IntGrid values.

LoadingStateAppExt

Extension trait for Bevy Apps to add loading states idiomatically

MaterialTilemap

PhysicsLayer

A layer used for determining which entities should interact with each other. Physics layers are used heavily by CollisionLayers.

Plugin

A collection of Bevy app logic and configuration.

PluginGroup

Combines multiple Plugins into a single unit.

PositionConstraint

A positional constraint applies a positional correction with a given direction and magnitude at the local contact points r1 and r2.

ProgressSystem

Extension trait for systems with progress tracking

ReadOnlySystem

System types that do not modify the World when run. This is implemented for any systems whose parameters all implement ReadOnlySystemParam.

Reflect

The core trait of bevy_reflect, used for accessing and modifying data dynamically.

Resource

A type that can be inserted into a World as a singleton.

Rng

An automatically-implemented extension trait on RngCore providing high-level generic methods for sampling values and other convenience methods.

RngCore

The core of a random number generator.

SeedableRng

A random number generator that can be explicitly seeded.

Serialize

A data structure that can be serialized into any data format supported by Serde.

SliceRandom

Extension trait on slices, providing random mutation and sampling methods.

States

Types that can define world-wide states in a finite-state machine.

Struct

A trait used to power struct-like operations via reflection.

System

An ECS system that can be added to a Schedule

SystemParamFunction

A trait implemented for all functions that can be used as Systems.

SystemSet

Types that identify logical groups of systems.

TransformPoint

A trait for point transformation methods.

TupleStruct

A trait used to power tuple struct-like operations via reflection.

XpbdConstraint

A trait for all XPBD constraints.

Functions

any_component_removed

Generates a Condition-satisfying closure that returns true if there are any entity with a component of the given type removed.

any_with_component

Generates a Condition-satisfying closure that returns true if there are any entities with the given component type.

apply_deferred

Instructs the executor to call System::apply_deferred on the systems that have run but not applied their Deferred system parameters (like Commands) or other system buffers.

apply_state_transition

If a new state is queued in NextState<S>, this system:

chunk_aabb

Calculates the Aabb of a generic chunk. The AABB depends upon the grid size, tile size, and map type of the the chunk the tile belongs to.

chunk_index_to_world_space

Calculates the world-space position of the bottom-left of the specified chunk.

dbg

System adapter that utilizes the bevy_utils::tracing::debug! macro to print the output of a system.

default

An ergonomic abbreviation for Default::default() to make initializing structs easier. This is especially helpful when combined with “struct update syntax”.

despawn_all

Convenience system for despawning all entities that match a given query filter

despawn_all_recursive

despawn_with_children_recursive

Function for despawning an entity and all its children

error

System adapter that utilizes the bevy_utils::tracing::error! macro to print the output of a system.

fill_tilemap

Fills an entire tile storage with the given tile.

fill_tilemap_hexagon

Fills a hexagonal region with the given tile_texture.

fill_tilemap_rect

Fills a rectangular region with the given tile.

fill_tilemap_rect_color

Fills a rectangular region with colored versions of the given tile.

generate_hex_ring

Generates a vector of hex positions that form a ring of given radius around the specified origin.

generate_hexagon

Generates a vector of hex positions that form a hexagon of given radius around the specified origin.

get_tilemap_center_transform

Calculates a Transform for a tilemap that places it so that its center is at (0.0, 0.0, 0.0) in world space.

ignore

System adapter that ignores the output of the previous system in a pipe. This is useful for fallible systems that should simply return early in case of an Err/None.

in_state

Generates a Condition-satisfying closure that returns true if the state machine is currently in state.

info

System adapter that utilizes the bevy_utils::tracing::info! macro to print system information.

init_resource

Convenience system for initting a resource of the given type

not

Generates a Condition that inverses the result of passed one.

on_event

Generates a Condition-satisfying closure that returns true if there are any new events of the given type since it was last called.

pause

Pause the simulation.

random

Generates a random value using the thread-local random number generator.

remove_from_all

Convenience system for removing a component from all entities that match a given query filter

remove_resource

Convenience system for removing a resource of the given type

resource_added

Generates a Condition-satisfying closure that returns true if the resource of the given type has been added since the condition was last checked.

resource_changed

Generates a Condition-satisfying closure that returns true if the resource of the given type has had its value changed since the condition was last checked.

resource_changed_or_removed

Generates a Condition-satisfying closure that returns true if the resource of the given type has had its value changed since the condition was last checked.

resource_equals

Generates a Condition-satisfying closure that returns true if the resource is equal to value.

resource_exists

Generates a Condition-satisfying closure that returns true if the resource exists.

resource_exists_and_changed

Generates a Condition-satisfying closure that returns true if the resource of the given type has had its value changed since the condition was last checked.

resource_exists_and_equals

Generates a Condition-satisfying closure that returns true if the resource exists and is equal to value.

resource_removed

Generates a Condition-satisfying closure that returns true if the resource of the given type has been removed since the condition was last checked.

resume

Resume the simulation.

run_once

Generates a Condition-satisfying closure that returns true if the first time the condition is run and false every time after

solve_constraint

Iterates through the constraints of a given type and solves them. Sleeping bodies are woken up when active bodies interact with them in a constraint.

state_changed

Generates a Condition-satisfying closure that returns true if the state machine changed state.

state_exists

Generates a Condition-satisfying closure that returns true if the state machine exists.

state_exists_and_equals

Generates a Condition-satisfying closure that returns true if the state machine exists and is currently in state.

thread_rng

Retrieve the lazily-initialized thread-local random number generator, seeded by the system. Intended to be used in method chaining style, e.g. thread_rng().gen::<i32>(), or cached locally, e.g. let mut rng = thread_rng();. Invoked by the Default trait, making ThreadRng::default() equivalent.

unwrap

System adapter that unwraps the Ok variant of a Result. This is useful for fallible systems that should panic in the case of an error.

warn

System adapter that utilizes the bevy_utils::tracing::warn! macro to print the output of a system.

Type Definitions

AnyResult

Result<T, Error>

AudioBundle

Bundle for playing a standard bevy audio asset

ColorMesh2dBundle

A component bundle for entities with a Mesh2dHandle and a ColorMaterial.

HashMap

A HashMap implementing aHash, a high speed keyed hashing algorithm intended for use in in-memory hashmaps.

HashSet

A HashSet implementing aHash, a high speed keyed hashing algorithm intended for use in in-memory hashmaps.

Instant

SpatialAudioBundle

Bundle for playing a standard bevy audio asset with a 3D position

TilemapBundle

Attribute Macros

bevy_main

reflect_trait

A macro that automatically generates type data for traits, which their implementors can then register.

serde_as

Convenience macro to use the serde_as system.

Derive Macros

AssetCollection

Derive macro for AssetCollection

Bundle

Component

Deref

Implements Deref for structs. This is especially useful when utilizing the newtype pattern.

DerefMut

Implements DerefMut for structs. This is especially useful when utilizing the newtype pattern.

Deserialize

DeserializeFromStr

Deserialize value by using its FromStr implementation

DynamicPlugin

Generates a dynamic plugin entry point function for the given Plugin type.

Error

Event

FromReflect

Derives the FromReflect trait.

LdtkEntity

LdtkIntCell

PhysicsLayer

Reflect

The main derive macro used by bevy_reflect for deriving its Reflect trait.

Resource

Serialize

SerializeDisplay

Serialize value by using it’s Display implementation

States

SystemSet

Derive macro generating an impl of the trait SystemSet.