Trait theseeker_engine::prelude::collision::AnyCollider

pub trait AnyCollider: Component {
    // Required methods
    fn aabb(
        &self,
        position: Vec2,
        rotation: impl Into<Rotation>
    ) -> ColliderAabb;
    fn mass_properties(&self, density: f32) -> ColliderMassProperties;
    fn contact_manifolds(
        &self,
        other: &Self,
        position1: Vec2,
        rotation1: impl Into<Rotation>,
        position2: Vec2,
        rotation2: impl Into<Rotation>,
        prediction_distance: f32
    ) -> Vec<ContactManifold>;

    // Provided method
    fn swept_aabb(
        &self,
        start_position: Vec2,
        start_rotation: impl Into<Rotation>,
        end_position: Vec2,
        end_rotation: impl Into<Rotation>
    ) -> ColliderAabb { ... }
}

A trait that generalizes over colliders. Implementing this trait allows colliders to be used with the physics engine.

Required Methods

fn aabb(&self, position: Vec2, rotation: impl Into<Rotation>) -> ColliderAabb

Computes the Axis-Aligned Bounding Box of the collider with the given position and rotation.

The rotation is counterclockwise and in radians.

fn mass_properties(&self, density: f32) -> ColliderMassProperties

Computes the collider’s mass properties based on its shape and a given density.

fn contact_manifolds( &self, other: &Self, position1: Vec2, rotation1: impl Into<Rotation>, position2: Vec2, rotation2: impl Into<Rotation>, prediction_distance: f32 ) -> Vec<ContactManifold>

Computes all ContactManifolds between two colliders.

Returns an empty vector if the colliders are separated by a distance greater than prediction_distance or if the given shapes are invalid.

Provided Methods

fn swept_aabb( &self, start_position: Vec2, start_rotation: impl Into<Rotation>, end_position: Vec2, end_rotation: impl Into<Rotation> ) -> ColliderAabb

Computes the swept Axis-Aligned Bounding Box of the collider. This corresponds to the space the shape would occupy if it moved from the given start position to the given end position.

The rotation is counterclockwise and in radians.

Object Safety

This trait is not object safe.