Type Alias theseeker_engine::prelude::collision::VHACDParameters

pub type VHACDParameters = VHACDParameters;

Parameters controlling the VHACD convex decomposition algorithm.

See https://github.com/Unity-Technologies/VHACD#parameters for details.

Aliased Type

struct VHACDParameters {
    pub concavity: f32,
    pub alpha: f32,
    pub beta: f32,
    pub resolution: u32,
    pub plane_downsampling: u32,
    pub convex_hull_downsampling: u32,
    pub fill_mode: FillMode,
    pub convex_hull_approximation: bool,
    pub max_convex_hulls: u32,
}

Fields
concavity: f32

Maximum concavity.

Default: 0.1 (in 2D), 0.01 (in 3D). Valid range [0.0, 1.0].

alpha: f32

Controls the bias toward clipping along symmetry planes.

Default: 0.05. Valid Range: [0.0, 1.0].

beta: f32

Controls the bias toward clipping along revolution planes.

Default: 0.05. Valid Range: [0.0, 1.0].

resolution: u32

Resolution used during the voxelization stage.

Default: 256 (in 2D), 64 (in 3D).

plane_downsampling: u32

Controls the granularity of the search for the best clipping plane during the decomposition.

Default: 4

convex_hull_downsampling: u32

Controls the precision of the convex-hull generation process during the clipping plane selection stage.

Default: 4

fill_mode: FillMode

Controls the way the input mesh or polyline is being voxelized.

Default: FillMode::FloodFill { detect_cavities: false, detect_self_intersections: false }

convex_hull_approximation: bool

Controls whether the convex-hull should be approximated during the decomposition stage. Setting this to true increases performances with a slight degradation of the decomposition quality.

Default: true

max_convex_hulls: u32

Controls the max number of convex-hull generated by the convex decomposition.

Default: 1024