Struct theseeker_engine::prelude::debug::PhysicsGizmos

pub struct PhysicsGizmos {

    pub axis_lengths: Option<Vec2>,
    pub aabb_color: Option<Color>,
    pub collider_color: Option<Color>,
    pub sleeping_color_multiplier: Option<[f32; 4]>,
    pub contact_point_color: Option<Color>,
    pub contact_normal_color: Option<Color>,
    pub contact_normal_scale: ContactGizmoScale,
    pub joint_anchor_color: Option<Color>,
    pub joint_separation_color: Option<Color>,
    pub raycast_color: Option<Color>,
    pub raycast_point_color: Option<Color>,
    pub raycast_normal_color: Option<Color>,
    pub shapecast_color: Option<Color>,
    pub shapecast_shape_color: Option<Color>,
    pub shapecast_point_color: Option<Color>,
    pub shapecast_normal_color: Option<Color>,
    pub hide_meshes: bool,

}

Gizmos used for debug rendering physics. See PhysicsDebugPlugin

You can configure what is rendered by inserting the PhysicsGizmos configuration group or by getting it from the GizmoConfigStore resource and mutating the returned configuration.

To configure the debug rendering of specific entities, use the DebugRender component.

Example

fn main() {
    App::new()
        .add_plugins((
            DefaultPlugins,
            PhysicsPlugins::default(),
            // Enables debug rendering
            PhysicsDebugPlugin::default(),
        ))
        // Overwrite default debug rendering configuration (optional)
        .insert_gizmo_group(
            PhysicsGizmos {
                aabb_color: Some(Color::WHITE),
                ..default()
            },
            GizmoConfig::default(),
        )
        .run();
}

fn setup(mut commands: Commands) {
    // This rigid body and its collider and AABB will get rendered
    commands.spawn((RigidBody::Dynamic, Collider::circle(0.5)));
}

Fields
axis_lengths: Option<Vec2>

The lengths of the axes drawn for an entity at the center of mass.

aabb_color: Option<Color>

The color of the AABBs. If None, the AABBs will not be rendered.

collider_color: Option<Color>

The color of the collider wireframes. If None, the colliders will not be rendered.

sleeping_color_multiplier: Option<[f32; 4]>

The colors (in HSLA) for sleeping bodies will be multiplied by this array. If None, sleeping will have no effect on the colors.

contact_point_color: Option<Color>

The color of the contact points. If None, the contact points will not be rendered.

contact_normal_color: Option<Color>

The color of the contact normals. If None, the contact normals will not be rendered.

contact_normal_scale: ContactGizmoScale

The scale used for contact normals.

joint_anchor_color: Option<Color>

The color of the lines drawn from the centers of bodies to their joint anchors.

joint_separation_color: Option<Color>

The color of the lines drawn between joint anchors, indicating the separation.

raycast_color: Option<Color>

The color used for the rays in raycasts.

raycast_point_color: Option<Color>

The color used for the hit points in raycasts.

raycast_normal_color: Option<Color>

The color used for the hit normals in raycasts.

shapecast_color: Option<Color>

The color used for the ray in shapecasts.

shapecast_shape_color: Option<Color>

The color used for the shape in shapecasts.

shapecast_point_color: Option<Color>

The color used for the hit points in shapecasts.

shapecast_normal_color: Option<Color>

The color used for the hit normals in shapecasts.

hide_meshes: bool

Determines if the visibility of entities with colliders should be set to Visibility::Hidden, which will only show the debug renders.

Implementations

impl PhysicsGizmos

pub fn all() -> PhysicsGizmos

Creates a PhysicsGizmos configuration with all rendering options enabled.

pub fn none() -> PhysicsGizmos

Creates a PhysicsGizmos configuration with debug rendering enabled but all options turned off.

Note: this doesn’t affect entities with DebugRender component; their debug gizmos will be visible.

pub fn axes(axis_lengths: Vec2) -> PhysicsGizmos

Creates a PhysicsGizmos configuration with the given lengths for the axes that are drawn for the entity at the center of mass. Other debug rendering options will be disabled.

pub fn aabbs(color: Color) -> PhysicsGizmos

Creates a PhysicsGizmos configuration with a given AABB color. Other debug rendering options will be disabled.

pub fn colliders(color: Color) -> PhysicsGizmos

Creates a PhysicsGizmos configuration with a given collider color. Other debug rendering options will be disabled.

pub fn contact_points(color: Color) -> PhysicsGizmos

Creates a PhysicsGizmos configuration with a given contact point color. Other debug rendering options will be disabled.

pub fn contact_normals(color: Color) -> PhysicsGizmos

Creates a PhysicsGizmos configuration with a given contact normal color. Other debug rendering options will be disabled.

pub fn joints( anchor_color: Option<Color>, separation_color: Option<Color> ) -> PhysicsGizmos

Creates a PhysicsGizmos configuration with given colors for joint anchors and separation distances. Other debug rendering options will be disabled.

pub fn with_axes(self, axis_lengths: Vec2) -> PhysicsGizmos

Sets the lengths of the axes drawn for the entity.

pub fn with_aabb_color(self, color: Color) -> PhysicsGizmos

Sets the AABB color.

pub fn with_collider_color(self, color: Color) -> PhysicsGizmos

Sets the collider color.

pub fn with_sleeping_color_multiplier( self, color_multiplier: [f32; 4] ) -> PhysicsGizmos

Sets the multiplier used for the colors (in HSLA) of sleeping bodies.

pub fn with_contact_point_color(self, color: Color) -> PhysicsGizmos

Sets the contact point color.

pub fn with_contact_normal_color(self, color: Color) -> PhysicsGizmos

Sets the contact normal color.

pub fn with_contact_normal_scale( self, scale: ContactGizmoScale ) -> PhysicsGizmos

Sets the contact normal scale.

pub fn with_joint_colors( anchor_color: Option<Color>, separation_color: Option<Color> ) -> PhysicsGizmos

Sets the colors used for debug rendering joints.

pub fn with_raycast_colors( self, ray: Option<Color>, hit_point: Option<Color>, hit_normal: Option<Color> ) -> PhysicsGizmos

Sets the colors used for debug rendering raycasts.

pub fn with_shapecast_colors( self, ray: Option<Color>, shape: Option<Color>, hit_point: Option<Color>, hit_normal: Option<Color> ) -> PhysicsGizmos

Sets the colors used for debug rendering shapecasts.

pub fn with_mesh_visibility(self, is_visible: bool) -> PhysicsGizmos

Sets the visibility of the entity’s visual mesh.

pub fn without_axes(self) -> PhysicsGizmos

Disables axis debug rendering.

pub fn without_aabbs(self) -> PhysicsGizmos

Disables AABB debug rendering.

pub fn without_colliders(self) -> PhysicsGizmos

Disables collider debug rendering.

pub fn without_contact_points(self) -> PhysicsGizmos

Disables contact point debug rendering.

pub fn without_contact_normals(self) -> PhysicsGizmos

Disables contact normal debug rendering.

pub fn without_joints(self) -> PhysicsGizmos

Disables joint debug rendering.

pub fn without_raycasts(self) -> PhysicsGizmos

Disables raycast debug rendering.

pub fn without_shapecasts(self) -> PhysicsGizmos

Disables shapecast debug rendering.