Trait theseeker_engine::prelude::debug::PhysicsGizmoExt

pub trait PhysicsGizmoExt {
    // Required methods
    fn draw_line(&mut self, a: Vec2, b: Vec2, color: Color);
    fn draw_line_strip(
        &mut self,
        points: Vec<Vec2>,
        position: impl Into<Position>,
        rotation: impl Into<Rotation>,
        closed: bool,
        color: Color
    );
    fn draw_polyline(
        &mut self,
        vertices: &[Vec2],
        indices: &[[u32; 2]],
        position: impl Into<Position>,
        rotation: impl Into<Rotation>,
        color: Color
    );
    fn draw_arrow(&mut self, a: Vec2, b: Vec2, head_length: f32, color: Color);
    fn draw_collider(
        &mut self,
        collider: &Collider,
        position: impl Into<Position>,
        rotation: impl Into<Rotation>,
        color: Color
    );
    fn draw_raycast(
        &mut self,
        origin: Vec2,
        direction: Direction2d,
        max_time_of_impact: f32,
        hits: &[RayHitData],
        ray_color: Color,
        point_color: Color,
        normal_color: Color
    );
    fn draw_shapecast(
        &mut self,
        shape: &Collider,
        origin: Vec2,
        shape_rotation: impl Into<Rotation>,
        direction: Direction2d,
        max_time_of_impact: f32,
        hits: &[ShapeHitData],
        ray_color: Color,
        shape_color: Color,
        point_color: Color,
        normal_color: Color
    );
}

An extension trait for Gizmos<PhysicsGizmo>.

Required Methods

fn draw_line(&mut self, a: Vec2, b: Vec2, color: Color)

Draws a line from a to b.

fn draw_line_strip( &mut self, points: Vec<Vec2>, position: impl Into<Position>, rotation: impl Into<Rotation>, closed: bool, color: Color )

Draws lines between a list of points.

fn draw_polyline( &mut self, vertices: &[Vec2], indices: &[[u32; 2]], position: impl Into<Position>, rotation: impl Into<Rotation>, color: Color )

Draws a polyline based on the given vertex and index buffers.

fn draw_arrow(&mut self, a: Vec2, b: Vec2, head_length: f32, color: Color)

Draws an arrow from a to b with an arrowhead that has a length of head_length.

fn draw_collider( &mut self, collider: &Collider, position: impl Into<Position>, rotation: impl Into<Rotation>, color: Color )

Draws a Collider shape.

fn draw_raycast( &mut self, origin: Vec2, direction: Direction2d, max_time_of_impact: f32, hits: &[RayHitData], ray_color: Color, point_color: Color, normal_color: Color )

Draws the results of a raycast.

fn draw_shapecast( &mut self, shape: &Collider, origin: Vec2, shape_rotation: impl Into<Rotation>, direction: Direction2d, max_time_of_impact: f32, hits: &[ShapeHitData], ray_color: Color, shape_color: Color, point_color: Color, normal_color: Color )

Draws the results of a shapecast.

Object Safety

This trait is not object safe.