Module theseeker_engine::prelude::helpers::hex_grid::axial

Code for the axial coordinate system.

Structs

AxialPos

A position in a hex grid labelled according to HexCoordSystem::Row or HexCoordSystem::Column. It is composed of a pair of i32 digits named q and r. When converting from a TilePos, TilePos.x is mapped to q, while TilePos.y is mapped to r.

FractionalAxialPos

A fractional axial position can represent a point that lies inside a hexagon. It is typically the result of mapping a world position into hexagonal space.

Constants

COL_BASIS

The matrix for mapping from AxialPos, to world position when hexes are arranged in column format (“flat top” per Red Blob Games). See Size and Spacing at Red Blob Games for an interactive visual explanation, but note that: 1) we consider increasing-y to be the same as “going up”, while RBG considers increasing-y to be “going down”, 2) our vectors have magnitude 1 (in order to allow for easy scaling based on grid-size)

INV_COL_BASIS

The inverse of COL_BASIS.

INV_ROW_BASIS

The inverse of ROW_BASIS.

ROW_BASIS

The matrix for mapping from AxialPos, to world position when hexes are arranged in row format (“pointy top” per Red Blob Games). See Size and Spacing at Red Blob Games for an interactive visual explanation, but note that: 1) we consider increasing-y to be the same as “going up”, while RBG considers increasing-y to be “going down”, 2) our vectors have magnitude 1 (in order to allow for easy scaling based on grid-size)

UNIT_Q

UNIT_R

UNIT_S