Enum theseeker_engine::prelude::helpers::hex_grid::neighbors::HexDirection

pub enum HexDirection {
    Zero,
    One,
    Two,
    Three,
    Four,
    Five,
}

Neighbors of a hexagonal tile. Zero corresponds with East for row-oriented tiles, and North for column-oriented tiles. It might also correspond with a custom direction for hex maps which are oriented by a non-standard rotation.

Mathematically: Zero corresponds with “zero multiples of pi/3”, One with “one multiple of pi/3”, and so on.

HexDirections can be added, and subtracted (under the hood, it is addition/subtraction modulo 6).

HexDirections can be converted from/into usize, u32, isize, i32.

Variants

Zero

One

Two

Three

Four

Five

Implementations

impl HexDirection

pub fn offset(&self, tile_pos: &TilePos, coord_sys: HexCoordSystem) -> TilePos