Struct theseeker_engine::prelude::helpers::hex_grid::neighbors::HexNeighbors

pub struct HexNeighbors<T> {
    pub zero: Option<T>,
    pub one: Option<T>,
    pub two: Option<T>,
    pub three: Option<T>,
    pub four: Option<T>,
    pub five: Option<T>,
}

Stores some data T associated with each neighboring hex cell, if present.

Fields
zero: Option<T>one: Option<T>two: Option<T>three: Option<T>four: Option<T>five: Option<T>
Implementations

impl<T> HexNeighbors<T>

pub fn get(&self, direction: HexDirection) -> Option<&T>

Get an item that lies in a particular hex direction, specified by a HexDirection.

Will be None if no such items exists.

pub fn get_inner_mut(&mut self, direction: HexDirection) -> Option<&mut T>

Get a mutable reference to an item that lies in a particular hex direction.

Will be None if no such items exists.

pub fn get_mut(&mut self, direction: HexDirection) -> &mut Option<T>

Get a mutable reference to the optional item that lies in a particular hex direction.

Will be None if no such items exists.

pub fn set(&mut self, direction: HexDirection, data: T)

Set item that lies in a particular hex direction.

This does an Option::replace under the hood.

pub fn iter(&self) -> impl Iterator<Item = &T>

Iterate over neighbors, in the order specified by HEX_DIRECTIONS.

If a neighbor is None, this iterator will skip it.

pub fn and_then<U, F>(self, f: F) -> HexNeighbors<U>
where F: Fn(T) -> Option<U>,

Applies the supplied closure f with an and_then to each neighbor element, where f takes T by value.

pub fn and_then_ref<'a, U, F>(&'a self, f: F) -> HexNeighbors<U>
where F: Fn(&'a T) -> Option<U>,

Applies the supplied closure f with an and_then to each neighbor element, where f takes T by reference.

pub fn map_ref<'a, U, F>(&'a self, f: F) -> HexNeighbors<U>
where F: Fn(&'a T) -> U,

Applies the supplied closure f with a map to each neighbor element, where f takes T by reference.

pub fn from_directional_closure<F>(f: F) -> HexNeighbors<T>
where F: Fn(HexDirection) -> Option<T>,

Generates HexNeighbors<T> from a closure that takes a hex direction and outputs Option<T>.

impl HexNeighbors<TilePos>

pub fn get_neighboring_positions( tile_pos: &TilePos, map_size: &TilemapSize, hex_coord_sys: &HexCoordSystem ) -> HexNeighbors<TilePos>

Returns neighboring tile positions, given a coordinate system.

In general, if you know which coordinate system you are using, it will be more efficient to use one of: * HexNeighbors::get_neighboring_positions_standard * HexNeighbors::get_neighboring_positions_row_even * HexNeighbors::get_neighboring_positions_row_odd * HexNeighbors::get_neighboring_positions_col_even * HexNeighbors::get_neighboring_positions_col_odd

A tile position will be None for a particular direction, if that neighbor would not lie on the map.

pub fn get_neighboring_positions_standard( tile_pos: &TilePos, map_size: &TilemapSize ) -> HexNeighbors<TilePos>

Returns neighboring tile positions. This works for maps using HexCoordSystem::Row and HexCoordSystem::Column.

For maps using HexCoordSystem::RowEven, HexCoordSystem::ColumnEven, HexCoordSystem::RowOdd, HexCoordSystem::RowOdd, use one of: * HexNeighbors::get_neighboring_positions_row_even * HexNeighbors::get_neighboring_positions_row_odd * HexNeighbors::get_neighboring_positions_col_even * HexNeighbors::get_neighboring_positions_col_odd (Or, just don’t use a map that with a odd/even coordinate system, and prefer to use one with HexCoordSystem::Row or HexCoordSystem::Column).

A tile position will be None for a particular direction, if that neighbor would not lie on the map.

pub fn get_neighboring_positions_row_even( tile_pos: &TilePos, map_size: &TilemapSize ) -> HexNeighbors<TilePos>

Returns neighboring tile positions on a map using HexCoordSystem::RowEven.

A tile position will be None for a particular direction, if that neighbor would not lie on the map.

pub fn get_neighboring_positions_row_odd( tile_pos: &TilePos, map_size: &TilemapSize ) -> HexNeighbors<TilePos>

Returns neighboring tile positions on a map using HexCoordSystem::RowOdd.

A tile position will be None for a particular direction, if that neighbor would not lie on the map.

pub fn get_neighboring_positions_col_even( tile_pos: &TilePos, map_size: &TilemapSize ) -> HexNeighbors<TilePos>

Returns neighboring tile positions on a map using HexCoordSystem::ColumnEven.

A tile position will be None for a particular direction, if that neighbor would not lie on the map.

pub fn get_neighboring_positions_col_odd( tile_pos: &TilePos, map_size: &TilemapSize ) -> HexNeighbors<TilePos>

Returns neighboring tile positions on a map using HexCoordSystem::ColumnOdd.

A tile position will be None for a particular direction, if that neighbor would not lie on the map.

pub fn entities(&self, tile_storage: &TileStorage) -> HexNeighbors<Entity>

Returns the entities associated with each tile position.