Struct theseeker_engine::prelude::ldtk::FieldDefinition

pub struct FieldDefinition {

    pub field_definition_type: String,
    pub accept_file_types: Option<Vec<String>>,
    pub allowed_refs: AllowedRefs,
    pub allowed_refs_entity_uid: Option<i32>,
    pub allowed_ref_tags: Vec<String>,
    pub allow_out_of_level_ref: bool,
    pub array_max_length: Option<i32>,
    pub array_min_length: Option<i32>,
    pub auto_chain_ref: bool,
    pub can_be_null: bool,
    pub default_override: Option<Value>,
    pub doc: Option<String>,
    pub editor_always_show: bool,
    pub editor_cut_long_values: bool,
    pub editor_display_color: Option<String>,
    pub editor_display_mode: EditorDisplayMode,
    pub editor_display_pos: EditorDisplayPos,
    pub editor_display_scale: f32,
    pub editor_link_style: EditorLinkStyle,
    pub editor_show_in_world: bool,
    pub editor_text_prefix: Option<String>,
    pub editor_text_suffix: Option<String>,
    pub export_to_toc: bool,
    pub identifier: String,
    pub is_array: bool,
    pub max: Option<f32>,
    pub min: Option<f32>,
    pub regex: Option<String>,
    pub searchable: bool,
    pub symmetrical_ref: bool,
    pub text_language_mode: Option<TextLanguageMode>,
    pub tileset_uid: Option<i32>,
    pub purple_type: String,
    pub uid: i32,
    pub use_for_smart_color: bool,

}

This section is mostly only intended for the LDtk editor app itself. You can safely ignore it.

Fields
field_definition_type: String

Human readable value type. Possible values: Int, Float, String, Bool, Color, ExternEnum.XXX, LocalEnum.XXX, Point, FilePath.
If the field is an array, this field will look like Array<...> (eg. Array<Int>, Array<Point> etc.)
NOTE: if you enable the advanced option Use Multilines type, you will have “Multilines” instead of “String” when relevant.

accept_file_types: Option<Vec<String>>

Optional list of accepted file extensions for FilePath value type. Includes the dot: .ext

allowed_refs: AllowedRefs

Possible values: Any, OnlySame, OnlyTags, OnlySpecificEntity

allowed_refs_entity_uid: Option<i32>allowed_ref_tags: Vec<String>allow_out_of_level_ref: boolarray_max_length: Option<i32>

Array max length

array_min_length: Option<i32>

Array min length

auto_chain_ref: boolcan_be_null: bool

TRUE if the value can be null. For arrays, TRUE means it can contain null values (exception: array of Points can’t have null values).

default_override: Option<Value>

Default value if selected value is null or invalid.

doc: Option<String>

User defined documentation for this field to provide help/tips to level designers about accepted values.

editor_always_show: booleditor_cut_long_values: booleditor_display_color: Option<String>editor_display_mode: EditorDisplayMode

Possible values: Hidden, ValueOnly, NameAndValue, EntityTile, LevelTile, Points, PointStar, PointPath, PointPathLoop, RadiusPx, RadiusGrid, ArrayCountWithLabel, ArrayCountNoLabel, RefLinkBetweenPivots, RefLinkBetweenCenters

editor_display_pos: EditorDisplayPos

Possible values: Above, Center, Beneath

editor_display_scale: f32editor_link_style: EditorLinkStyle

Possible values: ZigZag, StraightArrow, CurvedArrow, ArrowsLine, DashedLine

editor_show_in_world: booleditor_text_prefix: Option<String>editor_text_suffix: Option<String>export_to_toc: bool

If TRUE, the field value will be exported to the toc project JSON field. Only applies to Entity fields.

identifier: String

User defined unique identifier

is_array: bool

TRUE if the value is an array of multiple values

max: Option<f32>

Max limit for value, if applicable

min: Option<f32>

Min limit for value, if applicable

regex: Option<String>

Optional regular expression that needs to be matched to accept values. Expected format: /some_reg_ex/g, with optional “i” flag.

searchable: bool

If enabled, this field will be searchable through LDtk command palette

symmetrical_ref: booltext_language_mode: Option<TextLanguageMode>

Possible values: <null>, LangPython, LangRuby, LangJS, LangLua, LangC, LangHaxe, LangMarkdown, LangJson, LangXml, LangLog

tileset_uid: Option<i32>

UID of the tileset used for a Tile

purple_type: String

Internal enum representing the possible field types. Possible values: F_Int, F_Float, F_String, F_Text, F_Bool, F_Color, F_Enum(…), F_Point, F_Path, F_EntityRef, F_Tile

uid: i32

Unique Int identifier

use_for_smart_color: bool

If TRUE, the color associated with this field will override the Entity or Level default color in the editor UI. For Enum fields, this would be the color associated to their values.