theseeker_engine::prelude
Module shape

Generation for some primitive shape meshes.

Structs

BoxDeprecated

An axis-aligned box defined by its minimum and maximum point.

CapsuleDeprecated

A cylinder with hemispheres at the top and bottom

CircleDeprecated

A circle in the XY plane

CubeDeprecated

CylinderDeprecated

A cylinder which stands on the XZ plane

IcosphereDeprecated

A sphere made from a subdivided Icosahedron.

PlaneDeprecated

A square on the XZ plane centered at the origin.

QuadDeprecated

A rectangle on the XY plane centered at the origin.

RegularPolygonDeprecated

A regular polygon in the XY plane

TorusDeprecated

A torus (donut) shape.

UVSphereDeprecated

A sphere made of sectors and stacks.