Struct theseeker_engine::prelude::spatial_query::RayCaster

pub struct RayCaster {
    pub enabled: bool,
    pub origin: DVec2,
    pub direction: DVec2,
    pub max_time_of_impact: f64,
    pub max_hits: u32,
    pub solid: bool,
    pub query_filter: SpatialQueryFilter,
    /* private fields */
}

A component used for ray casting.

Ray casting is a type of spatial query that finds one or more hits between a ray and a set of colliders.

Each ray is defined by a local origin and a direction. The RayCaster will find each hit and add them to the RayHits component. Each hit has a time_of_impact property which refers to how long the ray travelled, i.e. the distance between the origin and the point of intersection.

The RayCaster is the easiest way to handle simple ray casts. If you want more control and don’t want to perform ray casts every frame, consider using the SpatialQuery system parameter.

Hit count and order

The results of a ray cast are in an arbitrary order by default. You can iterate over them in the order of time of impact with the RayHits::iter_sorted method.

You can configure the maximum amount of hits for a ray using max_hits. By default this is unbounded, so you will get all hits. When the number or complexity of colliders is large, this can be very expensive computationally. Set the value to whatever works best for your case.

Note that when there are more hits than max_hits, some hits will be missed. To guarantee that the closest hit is included, you should set max_hits to one or a value that is enough to contain all hits.

Example

use bevy::prelude::*;
use bevy_xpbd_3d::prelude::*;

fn setup(mut commands: Commands) {
    // Spawn a ray at the center going right
    commands.spawn(RayCaster::new(Vec3::ZERO, Vec3::X));
    // ...spawn colliders and other things
}

fn print_hits(query: Query<(&RayCaster, &RayHits)>) {
    for (ray, hits) in &query {
        // For the faster iterator that isn't sorted, use `.iter()`
        for hit in hits.iter_sorted() {
            println!(
                "Hit entity {:?} at {} with normal {}",
                hit.entity,
                ray.origin + ray.direction * hit.time_of_impact,
                hit.normal,
            );
        }
    }
}

Fields
enabled: bool

Controls if the ray caster is enabled.

origin: DVec2

The local origin of the ray relative to the Position and Rotation of the ray entity or its parent.

To get the global origin, use the global_origin method.

direction: DVec2

The local direction of the ray relative to the Rotation of the ray entity or its parent.

To get the global direction, use the global_direction method.

max_time_of_impact: f64

The maximum distance the ray can travel. By default this is infinite, so the ray will travel until all hits up to max_hits have been checked.

max_hits: u32

The maximum number of hits allowed.

When there are more hits than max_hits, some hits will be missed. To guarantee that the closest hit is included, you should set max_hits to one or a value that is enough to contain all hits.

solid: bool

Controls how the ray behaves when the ray origin is inside of a collider.

If solid is true, the point of intersection will be the ray origin itself.
If solid is false, the collider will be considered to have no interior, and the point of intersection will be at the collider shape’s boundary.

query_filter: SpatialQueryFilter

Rules that determine which colliders are taken into account in the query.

Implementations

impl RayCaster

pub fn new(origin: DVec2, direction: DVec2) -> RayCaster

Creates a new RayCaster with a given origin and direction.

pub fn with_origin(self, origin: DVec2) -> RayCaster

Sets the ray origin.

pub fn with_direction(self, direction: DVec2) -> RayCaster

Sets the ray direction.

pub fn with_solidness(self, solid: bool) -> RayCaster

Sets if the ray treats colliders as solid.

If solid is true, the point of intersection will be the ray origin itself.
If solid is false, the collider will be considered to have no interior, and the point of intersection will be at the collider shape’s boundary.

pub fn with_max_time_of_impact(self, max_time_of_impact: f64) -> RayCaster

Sets the maximum time of impact, i.e. the maximum distance that the ray is allowed to travel.

pub fn with_max_hits(self, max_hits: u32) -> RayCaster

Sets the maximum number of allowed hits.

pub fn with_query_filter(self, query_filter: SpatialQueryFilter) -> RayCaster

Sets the ray caster’s query filter that controls which colliders should be included or excluded by ray casts.

pub fn enable(&mut self)

Enables the RayCaster.

pub fn disable(&mut self)

Disables the RayCaster.

pub fn global_origin(&self) -> DVec2

Returns the global origin of the ray.

pub fn global_direction(&self) -> DVec2

Returns the global direction of the ray.