Struct theseeker_engine::prelude::CollidingEntities

pub struct CollidingEntities(pub HashSet<Entity, BuildHasherDefault<AHasher>, Global>);

Contains the entities that are colliding with an entity.

This component is automatically added for all entities with a Collider.

Example

use bevy::prelude::*;
use bevy_xpbd_3d::prelude::*;

fn my_system(query: Query<(Entity, &CollidingEntities)>) {
    for (entity, colliding_entities) in &query {
        println!(
            "{:?} is colliding with the following entities: {:?}",
            entity,
            colliding_entities
        );
    }
}

Tuple Fields
0: HashSet<Entity, BuildHasherDefault<AHasher>, Global>