Struct theseeker_engine::prelude::LdtkProject

pub struct LdtkProject { /* private fields */ }

Main asset for loading LDtk project data.

Accessing level data

This type provides many methods for accessing level data. The correct method for you will vary depending on whether or not you need “complete” level data, and if so, whether or not your project uses internal levels or external levels.

Raw vs loaded levels

There are a couple main flavors that level data can have - raw and loaded.

Raw levels don’t have any type guarantee that the level data is complete or incomplete. Level data may be incomplete and contain no layer instances if external levels are enabled. However, even in this case, a raw level is sufficient if you don’t need any layer data. Raw levels are represented by the Level type from LDtk. See RawLevelAccessor and LevelMetadataAccessor for some methods that access raw levels.

On the other hand, loaded levels are type-guaranteed to have complete level data. Loaded levels are represented by the LoadedLevel type. Methods for accessing loaded levels vary depending on if the levels are internal or external.

Accessing internal and external loaded levels

By default, LDtk stores level data inside the main project file. You have the option to store level data externally, where each level gets its own file. In this case, some of the level data remains available in the project file, but not layer data. See the previous section for more details.

Level data stored so differently on disk results in a similar difference when loaded in memory. In the external case, an entirely different asset type LdtkExternalLevel comes into play. So, methods for accessing loaded levels vary between the two cases.

If you know that your project uses internal levels, you can coerce it as a “standalone project”. To do this, use LdtkProject::as_standalone. With that, you can use these loaded_level accessors.

If you know that your project uses external levels, you can coerce it as a “parent project”. To do this, use [LdtkProject::as_parent]. You will also need the LdtkExternalLevel asset collection. With these, you can use these external_level accessors.

Implementations

impl LdtkProject

Auto-generated by derive_getters::Getters.

pub fn data(&self) -> &LdtkProjectData

LDtk json data and level metadata.

pub fn tileset_map(&self) -> &HashMap<i32, Handle<Image>>

Map from tileset uids to image handles for the loaded tileset.

pub fn int_grid_image_handle(&self) -> &Option<Handle<Image>>

Image used for rendering int grid colors.

impl LdtkProject

pub fn json_data(&self) -> &LdtkJson

Raw ldtk json data.

pub fn as_standalone(&self) -> &LdtkJsonWithMetadata<InternalLevels>

Unwrap as a LdtkJsonWithMetadata<InternalLevels>. For use on internal-levels ldtk projects only.

Panics

Panics if self.data() is not LdtkProjectData::Standalone. This shouldn’t occur if the project uses internal levels.