Struct theseeker_engine::prelude::ShapeCaster

pub struct ShapeCaster {
    pub enabled: bool,
    pub shape: Collider,
    pub origin: DVec2,
    pub shape_rotation: f64,
    pub direction: DVec2,
    pub max_time_of_impact: f64,
    pub max_hits: u32,
    pub ignore_origin_penetration: bool,
    pub ignore_self: bool,
    pub query_filter: SpatialQueryFilter,
    /* private fields */
}

A component used for shapecasting.

Shapecasting is a type of spatial query where a shape travels along a straight line and computes hits with colliders. This is often used to determine how far an object can move in a direction before it hits something.

Each shapecast is defined by a shape (a Collider), its local shape_rotation, a local origin and a local direction. The ShapeCaster will find each hit and add them to the ShapeHits component in the order of the time of impact.

Computing lots of hits can be expensive, especially against complex geometry, so the maximum number of hits is one by default. This can be configured through the max_hits property.

The ShapeCaster is the easiest way to handle simple shapecasting. If you want more control and don’t want to perform shapecasts on every frame, consider using the SpatialQuery system parameter.

Example

use bevy::prelude::*;
use bevy_xpbd_3d::prelude::*;

fn setup(mut commands: Commands) {
    // Spawn a shape caster with a ball shape moving right starting from the origin
    commands.spawn(ShapeCaster::new(
        Collider::ball(0.5),
        Vec3::ZERO,
        Quat::default(),
        Vec3::X
    ));
}

fn print_hits(query: Query<(&ShapeCaster, &ShapeHits)>) {
    for (shape_caster, hits) in &query {
        for hit in hits.iter() {
            println!("Hit entity {:?}", hit.entity);
        }
    }
}

Fields
enabled: bool

Controls if the shape caster is enabled.

shape: Collider

The shape being cast represented as a Collider.

origin: DVec2

The local origin of the shape relative to the Position and Rotation of the shape caster entity or its parent.

To get the global origin, use the global_origin method.

shape_rotation: f64

The local rotation of the shape being cast relative to the Rotation of the shape caster entity or its parent. Expressed in radians.

To get the global shape rotation, use the global_shape_rotation method.

direction: DVec2

The local direction of the shapecast relative to the Rotation of the shape caster entity or its parent.

To get the global direction, use the global_direction method.

max_time_of_impact: f64

The maximum distance the shape can travel. By default this is infinite, so the shape will travel until a hit is found.

max_hits: u32

The maximum number of hits allowed. By default this is one and only the first hit is returned.

ignore_origin_penetration: bool

Controls how the shapecast behaves when the shape is already penetrating a collider at the shape origin.

If set to true and the shape is being cast in a direction where it will eventually stop penetrating, the shapecast will not stop immediately, and will instead continue until another hit.
If set to false, the shapecast will stop immediately and return the hit. This is the default.

ignore_self: bool

If true, the shape caster ignores hits against its own Collider. This is the default.

query_filter: SpatialQueryFilter

Rules that determine which colliders are taken into account in the query.

Implementations

impl ShapeCaster

pub fn new( shape: Collider, origin: DVec2, shape_rotation: f64, direction: DVec2 ) -> ShapeCaster

Creates a new ShapeCaster with a given shape, origin, shape rotation and direction.

pub fn with_origin(self, origin: DVec2) -> ShapeCaster

Sets the ray origin.

pub fn with_direction(self, direction: DVec2) -> ShapeCaster

Sets the ray direction.

pub fn with_ignore_origin_penetration(self, ignore: bool) -> ShapeCaster

Controls how the shapecast behaves when the shape is already penetrating a collider at the shape origin.

If set to true and the shape is being cast in a direction where it will eventually stop penetrating, the shapecast will not stop immediately, and will instead continue until another hit.
If set to false, the shapecast will stop immediately and return the hit. This is the default.

pub fn with_ignore_self(self, ignore: bool) -> ShapeCaster

Sets if the shape caster should ignore hits against its own Collider. The default is true.

pub fn with_max_time_of_impact(self, max_time_of_impact: f64) -> ShapeCaster

Sets the maximum time of impact, i.e. the maximum distance that the ray is allowed to travel.

pub fn with_max_hits(self, max_hits: u32) -> ShapeCaster

Sets the maximum number of allowed hits.

pub fn with_query_filter(self, query_filter: SpatialQueryFilter) -> ShapeCaster

Sets the shape caster’s query filter that controls which colliders should be included or excluded by shapecasts.

pub fn enable(&mut self)

Enables the ShapeCaster.

pub fn disable(&mut self)

Disables the ShapeCaster.

pub fn global_origin(&self) -> DVec2

Returns the global origin of the ray.

pub fn global_shape_rotation(&self) -> f64

Returns the global rotation of the shape.

pub fn global_direction(&self) -> DVec2

Returns the global direction of the ray.