Struct theseeker_engine::prelude::SpatialQueryPipeline

pub struct SpatialQueryPipeline { /* private fields */ }

A resource for the spatial query pipeline.

The pipeline maintains a quaternary bounding volume hierarchy Qbvh of the world’s colliders as an acceleration structure for spatial queries.

Implementations

impl SpatialQueryPipeline

pub fn new() -> SpatialQueryPipeline

Creates a new SpatialQueryPipeline.

pub fn update<'a>( &mut self, colliders: impl Iterator<Item = (Entity, &'a Position, &'a Rotation, &'a Collider, Option<&'a CollisionLayers>)>, added_colliders: impl Iterator<Item = Entity> )

Updates the associated acceleration structures with a new set of entities.

pub fn cast_ray( &self, origin: DVec2, direction: DVec2, max_time_of_impact: f64, solid: bool, query_filter: SpatialQueryFilter ) -> Option<RayHitData>

Casts a ray and computes the closest hit with a collider. If there are no hits, None is returned.

Arguments

origin: Where the ray is cast from.

direction: What direction the ray is cast in.

max_time_of_impact: The maximum distance that the ray can travel.

solid: If true and the ray origin is inside of a collider, the hit point will be the ray origin itself. Otherwise, the collider will be treated as hollow, and the hit point will be at the collider’s boundary.

query_filter: A SpatialQueryFilter that determines which colliders are taken into account in the query.

See also: SpatialQuery::cast_ray

pub fn cast_ray_predicate( &self, origin: DVec2, direction: DVec2, max_time_of_impact: f64, solid: bool, query_filter: SpatialQueryFilter, predicate: &dyn Fn(Entity) -> bool ) -> Option<RayHitData>

Casts a ray and computes the closest hit with a collider. If there are no hits, None is returned.

Arguments

origin: Where the ray is cast from.

direction: What direction the ray is cast in.

max_time_of_impact: The maximum distance that the ray can travel.

solid: If true and the ray origin is inside of a collider, the hit point will be the ray origin itself. Otherwise, the collider will be treated as hollow, and the hit point will be at the collider’s boundary.

query_filter: A SpatialQueryFilter that determines which colliders are taken into account in the query.

predicate: A function with which the colliders are filtered. Given the Entity it should return false, if the entity should be ignored.

See also: SpatialQuery::cast_ray

pub fn ray_hits( &self, origin: DVec2, direction: DVec2, max_time_of_impact: f64, max_hits: u32, solid: bool, query_filter: SpatialQueryFilter ) -> Vec<RayHitData>

Casts a ray and computes all hits until max_hits is reached.

Note that the order of the results is not guaranteed, and if there are more hits than max_hits, some hits will be missed.

Arguments

origin: Where the ray is cast from.

direction: What direction the ray is cast in.

max_time_of_impact: The maximum distance that the ray can travel.

max_hits: The maximum number of hits. Additional hits will be missed.

solid: If true and the ray origin is inside of a collider, the hit point will be the ray origin itself. Otherwise, the collider will be treated as hollow, and the hit point will be at the collider’s boundary.

query_filter: A SpatialQueryFilter that determines which colliders are taken into account in the query.

See also: SpatialQuery::ray_hits

pub fn ray_hits_callback( &self, origin: DVec2, direction: DVec2, max_time_of_impact: f64, solid: bool, query_filter: SpatialQueryFilter, callback: impl FnMut(RayHitData) -> bool )

Casts a ray and computes all hits, calling the given callback for each hit. The raycast stops when callback returns false or all hits have been found.

Note that the order of the results is not guaranteed.

Arguments

origin: Where the ray is cast from.

direction: What direction the ray is cast in.

max_time_of_impact: The maximum distance that the ray can travel.

solid: If true and the ray origin is inside of a collider, the hit point will be the ray origin itself. Otherwise, the collider will be treated as hollow, and the hit point will be at the collider’s boundary.

query_filter: A SpatialQueryFilter that determines which colliders are taken into account in the query.

callback: A callback function called for each hit.

See also: SpatialQuery::ray_hits_callback

pub fn cast_shape( &self, shape: &Collider, origin: DVec2, shape_rotation: f64, direction: DVec2, max_time_of_impact: f64, ignore_origin_penetration: bool, query_filter: SpatialQueryFilter ) -> Option<ShapeHitData>

Casts a shape with a given rotation and computes the closest hit with a collider. If there are no hits, None is returned.

For a more ECS-based approach, consider using the ShapeCaster component instead.

Arguments

shape: The shape being cast represented as a Collider.

origin: Where the shape is cast from.

shape_rotation: The rotation of the shape being cast.

direction: What direction the shape is cast in.

max_time_of_impact: The maximum distance that the shape can travel.

ignore_origin_penetration: If true and the shape is already penetrating a collider at the shape origin, the hit will be ignored and only the next hit will be computed. Otherwise, the initial hit will be returned.

query_filter: A SpatialQueryFilter that determines which colliders are taken into account in the query.

See also: SpatialQuery::cast_shape

pub fn shape_hits( &self, shape: &Collider, origin: DVec2, shape_rotation: f64, direction: DVec2, max_time_of_impact: f64, max_hits: u32, ignore_origin_penetration: bool, query_filter: SpatialQueryFilter ) -> Vec<ShapeHitData>

Casts a shape with a given rotation and computes computes all hits in the order of the time of impact until max_hits is reached.

Arguments

shape: The shape being cast represented as a Collider.

origin: Where the shape is cast from.

shape_rotation: The rotation of the shape being cast.

direction: What direction the shape is cast in.

max_time_of_impact: The maximum distance that the shape can travel.

max_hits: The maximum number of hits. Additional hits will be missed.

ignore_origin_penetration: If true and the shape is already penetrating a collider at the shape origin, the hit will be ignored and only the next hit will be computed. Otherwise, the initial hit will be returned.

query_filter: A SpatialQueryFilter that determines which colliders are taken into account in the query.

callback: A callback function called for each hit.

See also: SpatialQuery::shape_hits

pub fn shape_hits_callback( &self, shape: &Collider, origin: DVec2, shape_rotation: f64, direction: DVec2, max_time_of_impact: f64, ignore_origin_penetration: bool, query_filter: SpatialQueryFilter, callback: impl FnMut(ShapeHitData) -> bool )

Casts a shape with a given rotation and computes computes all hits in the order of the time of impact, calling the given callback for each hit. The shapecast stops when callback returns false or all hits have been found.

Arguments

shape: The shape being cast represented as a Collider.

origin: Where the shape is cast from.

shape_rotation: The rotation of the shape being cast.

direction: What direction the shape is cast in.

max_time_of_impact: The maximum distance that the shape can travel.

ignore_origin_penetration: If true and the shape is already penetrating a collider at the shape origin, the hit will be ignored and only the next hit will be computed. Otherwise, the initial hit will be returned.

query_filter: A SpatialQueryFilter that determines which colliders are taken into account in the query.

callback: A callback function called for each hit.

See also: SpatialQuery::shape_hits_callback

pub fn project_point( &self, point: DVec2, solid: bool, query_filter: SpatialQueryFilter ) -> Option<PointProjection>

Finds the projection of a given point on the closest collider. If one isn’t found, None is returned.

Arguments

point: The point that should be projected.

solid: If true and the point is inside of a collider, the projection will be at the point. Otherwise, the collider will be treated as hollow, and the projection will be at the collider’s boundary.

query_filter: A SpatialQueryFilter that determines which colliders are taken into account in the query.

See also: SpatialQuery::project_point

pub fn point_intersections( &self, point: DVec2, query_filter: SpatialQueryFilter ) -> Vec<Entity>

An intersection test that finds all entities with a collider that contains the given point.

Arguments

point: The point that intersections are tested against.

query_filter: A SpatialQueryFilter that determines which colliders are taken into account in the query.

See also: SpatialQuery::point_intersections

pub fn point_intersections_callback( &self, point: DVec2, query_filter: SpatialQueryFilter, callback: impl FnMut(Entity) -> bool )

An intersection test that finds all entities with a collider that contains the given point, calling the given callback for each intersection. The search stops when callback returns false or all intersections have been found.

Arguments

point: The point that intersections are tested against.

query_filter: A SpatialQueryFilter that determines which colliders are taken into account in the query.

callback: A callback function called for each intersection.

See also: SpatialQuery::point_intersections_callback

pub fn aabb_intersections_with_aabb(&self, aabb: ColliderAabb) -> Vec<Entity>

An intersection test that finds all entities with a ColliderAabb that is intersecting the given aabb.

See also: SpatialQuery::point_intersections_callback

pub fn aabb_intersections_with_aabb_callback( &self, aabb: ColliderAabb, callback: impl FnMut(Entity) -> bool )

An intersection test that finds all entities with a ColliderAabb that is intersecting the given aabb, calling callback for each intersection. The search stops when callback returns false or all intersections have been found.

See also: SpatialQuery::aabb_intersections_with_aabb_callback

pub fn shape_intersections( &self, shape: &Collider, shape_position: DVec2, shape_rotation: f64, query_filter: SpatialQueryFilter ) -> Vec<Entity>

An intersection test that finds all entities with a Collider that is intersecting the given shape with a given position and rotation.

Arguments

shape: The shape that intersections are tested against represented as a Collider.

shape_position: The position of the shape.

shape_rotation: The rotation of the shape.

query_filter: A SpatialQueryFilter that determines which colliders are taken into account in the query.

See also: SpatialQuery::shape_intersections

pub fn shape_intersections_callback( &self, shape: &Collider, shape_position: DVec2, shape_rotation: f64, query_filter: SpatialQueryFilter, callback: impl FnMut(Entity) -> bool )

An intersection test that finds all entities with a Collider that is intersecting the given shape with a given position and rotation, calling callback for each intersection. The search stops when callback returns false or all intersections have been found.

Arguments

shape: The shape that intersections are tested against represented as a Collider.

shape_position: The position of the shape.

shape_rotation: The rotation of the shape.

query_filter: A SpatialQueryFilter that determines which colliders are taken into account in the query.

callback: A callback function called for each intersection.

See also: SpatialQuery::shape_intersections_callback