Struct theseeker_engine::prelude::TextureAtlasLayout

pub struct TextureAtlasLayout {
    pub size: Vec2,
    pub textures: Vec<Rect>,
    /* private fields */
}

Stores a map used to lookup the position of a texture in a TextureAtlas. This can be used to either use and look up a specific section of a texture, or animate frame-by-frame as a sprite sheet.

Optionally it can store a mapping from sub texture handles to the related area index (see TextureAtlasBuilder).

Example usage animating sprite. Example usage loading sprite sheet.

Fields
size: Vec2textures: Vec<Rect>

The specific areas of the atlas where each texture can be found

Implementations

impl TextureAtlasLayout

pub fn new_empty(dimensions: Vec2) -> TextureAtlasLayout

Create a new empty layout with custom dimensions

pub fn from_grid( tile_size: Vec2, columns: usize, rows: usize, padding: Option<Vec2>, offset: Option<Vec2>, ) -> TextureAtlasLayout

Generate a TextureAtlasLayout as a grid where each tile_size by tile_size grid-cell is one of the section in the atlas. Grid cells are separated by some padding, and the grid starts at offset pixels from the top left corner. Resulting layout is indexed left to right, top to bottom.

Arguments

tile_size - Each layout grid cell size

columns - Grid column count

rows - Grid row count

padding - Optional padding between cells

offset - Optional global grid offset

pub fn add_texture(&mut self, rect: Rect) -> usize

Add a section to the list in the layout and returns its index which can be used with TextureAtlas

Arguments

rect - The section of the texture to be added

pub fn len(&self) -> usize

The number of textures in the TextureAtlasLayout

pub fn is_empty(&self) -> bool

pub fn get_texture_index( &self, texture: impl Into<AssetId<Image>>, ) -> Option<usize>

Retrieves the texture section index of the given texture handle.

This requires the layout to have been built using a TextureAtlasBuilder