Trait theseeker_engine::prelude::AngularConstraint

pub trait AngularConstraint: XpbdConstraint<2> {
    // Provided methods
    fn apply_angular_correction(
        &self,
        body1: &mut RigidBodyQueryItem<'_>,
        body2: &mut RigidBodyQueryItem<'_>,
        delta_lagrange: f32,
        axis: Vec3
    ) -> f32 { ... }
    fn compute_generalized_inverse_mass(
        &self,
        body: &RigidBodyQueryItem<'_>,
        axis: Vec3
    ) -> f32 { ... }
    fn get_delta_rot(_rot: Rotation, inverse_inertia: f32, p: f32) -> Rotation { ... }
    fn compute_torque(&self, lagrange: f32, axis: Vec3, dt: f32) -> f32 { ... }
}

An angular constraint applies an angular correction around a given axis.

Provided Methods

fn apply_angular_correction( &self, body1: &mut RigidBodyQueryItem<'_>, body2: &mut RigidBodyQueryItem<'_>, delta_lagrange: f32, axis: Vec3 ) -> f32

Applies angular constraints for interactions between two bodies.

Here in 2D, axis is a unit vector with the Z coordinate set to 1 or -1. It controls if the body should rotate counterclockwise or clockwise.

Returns the angular impulse that is applied proportional to the inverse masses of the bodies.

fn compute_generalized_inverse_mass( &self, body: &RigidBodyQueryItem<'_>, axis: Vec3 ) -> f32

Computes the generalized inverse mass of a body when applying an angular correction around axis.

In 2D, axis should only have the z axis set to either -1 or 1 to indicate counterclockwise or clockwise rotation.

fn get_delta_rot(_rot: Rotation, inverse_inertia: f32, p: f32) -> Rotation

Computes the update in rotation when applying an angular correction p.

fn compute_torque(&self, lagrange: f32, axis: Vec3, dt: f32) -> f32

Computes the torque acting along the constraint using the equation tau = lambda * n / h^2

Object Safety

This trait is not object safe.