theseeker_engine::prelude
Trait CliCommandsRegisterExt

pub trait CliCommandsRegisterExt {
    // Required methods
    fn register_clicommand_noargs<S, Param>(
        &mut self,
        name: &str,
        system: S,
    ) -> &mut Self
       where S: IntoSystem<(), (), Param>;
    fn register_clicommand_args<S, Param>(
        &mut self,
        name: &str,
        system: S,
    ) -> &mut Self
       where S: IntoSystem<Vec<String>, (), Param>;
    fn deregister_clicommand(&mut self, name: &str, has_args: bool) -> &mut Self;
}

Provides methods for managing the available “console commands”

There are two kinds of commands:

“args” (system takes Vec<String> as input)

“noargs” (system takes no input)

When you try to run a cli string, the correct variant will be chosen based on whether args were present.

You should be able to do this at App creation, as well as later at runtime, with exclusive World access.

Required Methods

fn register_clicommand_noargs<S, Param>( &mut self, name: &str, system: S, ) -> &mut Self
where S: IntoSystem<(), (), Param>,

Create a new “console command” with the given string and system/implementation

If a command with the same name already exists, it is replaced.

fn register_clicommand_args<S, Param>( &mut self, name: &str, system: S, ) -> &mut Self
where S: IntoSystem<Vec<String>, (), Param>,

Create a new “console command” with the given string and system/implementation

If a command with the same name already exists, it is replaced.

fn deregister_clicommand(&mut self, name: &str, has_args: bool) -> &mut Self

Remove a “console command”, if it exists

has_args specifies whether you want to remove the “args” or “noargs” variant.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.