theseeker_engine::prelude
Trait ConfigureLoadingState

pub trait ConfigureLoadingState {
    // Required methods
    fn load_collection<A>(self) -> Self
       where A: AssetCollection;
    fn init_resource<R>(self) -> Self
       where R: Resource + FromWorld;
    fn register_dynamic_asset_collection<C>(self) -> Self
       where C: DynamicAssetCollection + Asset;
    fn with_dynamic_assets_file<C>(self, file: &str) -> Self
       where C: DynamicAssetCollection + Asset;
}

Methods to configure a loading state

Required Methods

fn load_collection<A>(self) -> Self
where A: AssetCollection,

Add the given collection to the loading state.

Its loading progress will be tracked. Only when all included handles are fully loaded, the collection will be inserted to the ECS as a resource.

See the two_collections example

fn init_resource<R>(self) -> Self
where R: Resource + FromWorld,

The resource will be initialized at the end of the loading state using its FromWorld implementation. All asset collections will be available at that point and fully loaded.

See the init_resource example

fn register_dynamic_asset_collection<C>(self) -> Self
where C: DynamicAssetCollection + Asset,

Register a custom dynamic asset collection type

See the custom_dynamic_assets example

fn with_dynamic_assets_file<C>(self, file: &str) -> Self
where C: DynamicAssetCollection + Asset,

Add a file containing dynamic assets to the loading state. Keys contained in the file, will be available for asset collections.

See the dynamic_asset example

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.