theseeker_engine::prelude
Trait DynamicAssetCollection

pub trait DynamicAssetCollection {
    // Required method
    fn register(&self, dynamic_assets: &mut DynamicAssets);
}

This traits describes types that contain asset configurations and can register them in the DynamicAssets resource.

Required Methods

fn register(&self, dynamic_assets: &mut DynamicAssets)

Register all dynamic assets inside the collection in the DynamicAssets resource.