Trait theseeker_engine::prelude::IntoSystemAppConfigs

pub trait IntoSystemAppConfigs<Marker>: Sized {
    // Required method
    fn into_app_configs(self) -> SystemAppConfigs;

    // Provided methods
    fn in_schedule(self, label: impl ScheduleLabel) -> SystemAppConfigs { ... }
    fn on_startup(self) -> SystemAppConfigs { ... }
}

Types that can convert into SystemAppConfigs.

Required Methods

fn into_app_configs(self) -> SystemAppConfigs

Converts to SystemAppConfigs.

Provided Methods

fn in_schedule(self, label: impl ScheduleLabel) -> SystemAppConfigs

Adds the systems to the provided schedule.

If a schedule is not specified, they will be added to the App’s default schedule.

Panics

If any of the systems have already been assigned to a schedule.

fn on_startup(self) -> SystemAppConfigs

Adds the systems to CoreSchedule::Startup. This is a shorthand for self.in_schedule(CoreSchedule::Startup).

Examples

app.add_systems(
    (
        startup_system_a,
        startup_system_b,
        startup_system_c,
    )
        .on_startup()
);

Panics

If any of the systems have already been assigned to a schedule.