Trait theseeker_engine::prelude::LdtkEntityAppExt

pub trait LdtkEntityAppExt {
    // Required method
    fn register_ldtk_entity_for_layer_optional<B>(
        &mut self,
        layer_identifier: Option<String>,
        entity_identifier: Option<String>
    ) -> &mut Self
       where B: LdtkEntity + Bundle;

    // Provided methods
    fn register_ldtk_entity_for_layer<B>(
        &mut self,
        layer_identifier: &str,
        entity_identifier: &str
    ) -> &mut Self
       where B: LdtkEntity + Bundle { ... }
    fn register_ldtk_entity<B>(&mut self, entity_identifier: &str) -> &mut Self
       where B: LdtkEntity + Bundle { ... }
    fn register_default_ldtk_entity_for_layer<B>(
        &mut self,
        layer_identifier: &str
    ) -> &mut Self
       where B: LdtkEntity + Bundle { ... }
    fn register_default_ldtk_entity<B>(&mut self) -> &mut Self
       where B: LdtkEntity + Bundle { ... }
}

Provides functions to register Bundles to bevy’s App for particular LDtk layer and entity identifiers.

After being registered, entities will be spawned with these Bundles when some entity in LDtk meets the criteria you specify.

Not intended for custom implementations on your own types.

Required Methods

fn register_ldtk_entity_for_layer_optional<B>( &mut self, layer_identifier: Option<String>, entity_identifier: Option<String> ) -> &mut Selfwhere B: LdtkEntity + Bundle,

Used internally by all the other LDtk entity registration functions.

Similar to LdtkEntityAppExt::register_ldtk_entity_for_layer, except it provides defaulting functionality:

Setting layer_identifier to None will make the registration apply to any Entity layer.

Setting entity_identifier to None will make the registration apply to any LDtk entity.

This defaulting functionality means that a particular instance of an LDtk entity may match multiple registrations. In these cases, registrations are prioritized in order of most to least specific:

layer_identifier and entity_identifier are specified

Just entity_identifier is specified

Just layer_identifier is specified

Neither entity_identifier nor layer_identifier are specified

Provided Methods

fn register_ldtk_entity_for_layer<B>( &mut self, layer_identifier: &str, entity_identifier: &str ) -> &mut Selfwhere B: LdtkEntity + Bundle,

Registers LdtkEntity types to be spawned for a given Entity identifier and layer identifier in an LDtk file.

This example lets the plugin know that it should spawn a MyBundle when it encounters a “my_entity_identifier” entity on a “MyLayerIdentifier” layer.

use bevy::prelude::*;
use bevy_ecs_ldtk::prelude::*;

fn main() {
    App::empty()
        .add_plugin(LdtkPlugin)
        .register_ldtk_entity_for_layer::<MyBundle>("MyLayerIdentifier", "my_entity_identifier")
        // add other systems, plugins, resources...
        .run();
}

#[derive(Bundle, LdtkEntity)]
pub struct MyBundle {
    a: ComponentA,
    b: ComponentB,
    c: ComponentC,
}

You can find more details on the #[derive(LdtkEntity)] macro at LdtkEntity.

fn register_ldtk_entity<B>(&mut self, entity_identifier: &str) -> &mut Selfwhere B: LdtkEntity + Bundle,

Similar to LdtkEntityAppExt::register_ldtk_entity_for_layer, except it applies the registration to all layers.

fn register_default_ldtk_entity_for_layer<B>( &mut self, layer_identifier: &str ) -> &mut Selfwhere B: LdtkEntity + Bundle,

Similar to LdtkEntityAppExt::register_ldtk_entity_for_layer, except it applies the registration to all entities on the given layer.

fn register_default_ldtk_entity<B>(&mut self) -> &mut Selfwhere B: LdtkEntity + Bundle,

Similar to LdtkEntityAppExt::register_ldtk_entity_for_layer, except it applies the registration to any entity and any layer.