theseeker_engine::prelude
Trait MapKey

pub trait MapKey {
    // Required method
    fn from_asset_path(path: &AssetPath<'_>) -> Self;
}

A type that can be used as key for mapped asset collection.

String and Box<str>

Both String and Box<str> implements MapKey by using the path of the asset as the key.

Key collision

Following the implementation of the MapKey trait, key collisions may happen, resulting in some assets not being loaded. This is up to the user to ensure that there are no collisions.

Required Methods

fn from_asset_path(path: &AssetPath<'_>) -> Self

Creates the key from the path of the asset.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.