theseeker_engine::prelude
Trait ProgressSystem

pub trait ProgressSystem<Params, T>: IntoSystem<(), T, Params>
where
    T: ApplyProgress,
{
    // Required method
    fn track_progress(self) -> NodeConfigs<Box<dyn System<In = (), Out = ()>>>;
}

Extension trait for systems with progress tracking

Required Methods

fn track_progress(self) -> NodeConfigs<Box<dyn System<In = (), Out = ()>>>

Call this to add your system returning Progress to your App

This adds the functionality for tracking the returned Progress.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.