Trait theseeker_engine::prelude::RawLevelAccessor

pub trait RawLevelAccessor {
    // Required methods
    fn root_levels(&self) -> &[Level];
    fn worlds(&self) -> &[World];

    // Provided methods
    fn iter_root_levels(&self) -> Iter<'_, Level>  { ... }
    fn iter_world_levels(
        &self,
    ) -> FlatMap<Iter<'_, World>, Iter<'_, Level>, fn(_: &World) -> Iter<'_, Level>>  { ... }
    fn iter_raw_levels(
        &self,
    ) -> Chain<Iter<'_, Level>, FlatMap<Iter<'_, World>, Iter<'_, Level>, fn(_: &World) -> Iter<'_, Level>>>  { ... }
    fn iter_root_levels_with_indices(
        &self,
    ) -> Map<Enumerate<Iter<'_, Level>>, fn(_: (usize, &Level)) -> (LevelIndices, &Level)>  { ... }
    fn iter_world_levels_with_indices(
        &self,
    ) -> FlatMap<Enumerate<Iter<'_, World>>, Map<Zip<Repeat<usize>, Enumerate<Iter<'_, Level>>>, fn(_: (usize, (usize, &Level))) -> (LevelIndices, &Level)>, fn(_: (usize, &World)) -> Map<Zip<Repeat<usize>, Enumerate<Iter<'_, Level>>>, fn(_: (usize, (usize, &Level))) -> (LevelIndices, &Level)>>  { ... }
    fn iter_raw_levels_with_indices(
        &self,
    ) -> Chain<Map<Enumerate<Iter<'_, Level>>, fn(_: (usize, &Level)) -> (LevelIndices, &Level)>, FlatMap<Enumerate<Iter<'_, World>>, Map<Zip<Repeat<usize>, Enumerate<Iter<'_, Level>>>, fn(_: (usize, (usize, &Level))) -> (LevelIndices, &Level)>, fn(_: (usize, &World)) -> Map<Zip<Repeat<usize>, Enumerate<Iter<'_, Level>>>, fn(_: (usize, (usize, &Level))) -> (LevelIndices, &Level)>>>  { ... }
    fn get_raw_level_at_indices(&self, indices: &LevelIndices) -> Option<&Level> { ... }
}

Convenience methods for accessing raw level data by reference.

Root vs world levels

This trait is intended for types that store LdtkJson data. This sort of data stores levels both in the “root” of the project, and in each World. The former is referred to in trait methods as root_levels, and the latter as world_levels. Root levels may be removed in the future after LDtk’s multi-worlds update.

Required Methods

fn root_levels(&self) -> &[Level]

Slice to this project’s collection of root levels.

fn worlds(&self) -> &[World]

Slice to this project’s collection of Worlds.

Provided Methods

fn iter_root_levels(&self) -> Iter<'_, Level>

Iterate through this project’s root levels.

Note: all levels are considered raw.

fn iter_world_levels( &self, ) -> FlatMap<Iter<'_, World>, Iter<'_, Level>, fn(_: &World) -> Iter<'_, Level>>

Iterate through this project’s world levels.

Note: all levels are considered raw.

fn iter_raw_levels( &self, ) -> Chain<Iter<'_, Level>, FlatMap<Iter<'_, World>, Iter<'_, Level>, fn(_: &World) -> Iter<'_, Level>>>

Iterate through this project’s levels.

This first iterates through root levels, then world levels.

Note: all levels are considered raw.

fn iter_root_levels_with_indices( &self, ) -> Map<Enumerate<Iter<'_, Level>>, fn(_: (usize, &Level)) -> (LevelIndices, &Level)>

Iterate through this project’s root levels enumerated with their LevelIndices.

Note: all levels are considered raw.

fn iter_world_levels_with_indices( &self, ) -> FlatMap<Enumerate<Iter<'_, World>>, Map<Zip<Repeat<usize>, Enumerate<Iter<'_, Level>>>, fn(_: (usize, (usize, &Level))) -> (LevelIndices, &Level)>, fn(_: (usize, &World)) -> Map<Zip<Repeat<usize>, Enumerate<Iter<'_, Level>>>, fn(_: (usize, (usize, &Level))) -> (LevelIndices, &Level)>>

Iterate through this project’s world levels enumerated with their LevelIndices.

Note: all levels are considered raw.

fn iter_raw_levels_with_indices( &self, ) -> Chain<Map<Enumerate<Iter<'_, Level>>, fn(_: (usize, &Level)) -> (LevelIndices, &Level)>, FlatMap<Enumerate<Iter<'_, World>>, Map<Zip<Repeat<usize>, Enumerate<Iter<'_, Level>>>, fn(_: (usize, (usize, &Level))) -> (LevelIndices, &Level)>, fn(_: (usize, &World)) -> Map<Zip<Repeat<usize>, Enumerate<Iter<'_, Level>>>, fn(_: (usize, (usize, &Level))) -> (LevelIndices, &Level)>>>

Iterate through this project’s levels enumerated with their LevelIndices.

This first iterates through root levels, then world levels.

Note: all levels are considered raw.

fn get_raw_level_at_indices(&self, indices: &LevelIndices) -> Option<&Level>

Immutable access to a level at the given LevelIndices.

Note: all levels are considered raw.