Module theseeker_engine::assets::script

Structs

CommonScript

CommonScriptParams

CommonScriptSettings

ExtendedScript

ExtendedScriptParams

ExtendedScriptSettings

ExtendedScriptWorkaround

Flattened

Script

Scripted Sequence Asset type

ScriptConfig

When initting a script, quantize time (from TimeBase).

ScriptTickQuant

When initting a script, quantize time (from TimeBase).

Enums

CommonScriptAction

The various actions that can be performed from scripts

CommonScriptRunIf

ExtendedScriptAction

ExtendedScriptRunIf

ExtendedScriptWorkaroundInner

PlaybackControl

TimeBase

From what point does a script count time (when is time/tick 0)?