theseeker_engine
Module physics

Structs

AnimationCollider

An animation collider accepts a target entity that has the image, texture atlas, and sprite animation components.

Collider

Objects marked with this and a transform component will be updated in the collision scene. Parenting is not currently kept in sync; global transforms are used instead. Colliders ignore all scaling!

ColliderHandle

Just a wrapper that lets us treat rapiers collider handle as a component

LinearVelocity

A convenient component type for referring to velocity of an entity.

PhysicsPlugin

A manual implementation of rapier to only use the features required by our project

PhysicsSet

PhysicsWorld

Used to create queries on a physics world.

ShapeCaster

SpriteShapeMap

Constants

ENEMY

The enemy collision group

ENEMY_ATTACK

The enemy attack collision group

ENEMY_HURT

Combination of ENEMY and ENEMY_INSIDE, used for checking the players attacks

ENEMY_INSIDE

Applied to an enemy when player is inside it

GROUND

The ground collision group

PLAYER

The player collision group

PLAYER_ATTACK

The player attack collision group

SENSOR

Use when the other groups don’t make sense, and you just want to detect something

Functions

debug_colliders

Draws colliders using bevy’s gizmos to assist with debugging.

into_vec

Utility to convert from Vec2 to rapier compatible structure

into_vec2

Utility to convert from rapier type to Vec2

update_query_pipeline

Updates the pipeline by reading all the positions/components with colliders

update_sprite_colliders