Function theseeker_engine::physics::update_query_pipeline

pub fn update_query_pipeline(
    world: ResMut<'_, PhysicsWorld>,
    phys_obj_query: Query<'_, '_, (Entity, &Transform, Ref<'_, GlobalTransform>, Ref<'_, Collider>, Option<&ColliderHandle>)>,
    removed: RemovedComponents<'_, '_, Collider>,
    commands: Commands<'_, '_>
)

Updates the pipeline by reading all the positions/components with colliders

Make sure if you are reading from this in a system, you run after this finishes

TODO make sure this always runs after the GameTickUpdate; consider creating a seperate [ScheduleLabel] for immediately after transform propagation